### 2021-04-04 4316edd
remove gallery vignette
