+++ _pkgdown.yml
@@ -104,7 +104,6 @@ articles:
     - web_only/purrr_examples
     - web_only/principles
     - web_only/other_resources
-    - web_only/gallery
     - web_only/effsize_interpretation
     - web_only/session_info
 
