### 2021-04-05 #3141 @csoneson
👋 @RMHogervorst, @njtierney - would you be interested in reviewing this submission for JOSS?

> Visualizations with statistical details: The 'ggstatsplot' approach
