+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
 affiliations:
   - index: 1
   - name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
-date: "2021-03-30"
+date: "2021-04-05"
 bibliography: paper.bib
 ---
 
