+++ paper/paper.bib
@@ -101,10 +101,7 @@
     title = {{statsExpressions}: R Package for Tidy Dataframes and Expressions with Statistical Details},
     author = {Indrajeet Patil},
     year = {2021},
-    journal = {Journal of Open Source Software},
-    volume = {6},
-    number = {59},
-    pages = {3111},
+    journal = {PsyArXiv},
     doi = {10.31234/osf.io/ntbvy},
   }
 
@@ -150,10 +147,7 @@
     title = {Assessment, Testing and Comparison of Statistical Models using {R}},
     author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Philip Waggoner and Dominique Makowski},
     year = {2021},
-    journal = {Journal of Open Source Software},
-    volume = {6},
-    number = {59},
-    pages = {3112},
+    journal = {PsyArXiv},
     doi = {10.31234/osf.io/vtq8f},
   }
 
