### 2021-04-05 #3141 @IndrajeetPatil
@csoneson Sorry about that. I have [fixed](https://github.com/IndrajeetPatil/ggstatsplot/commit/94db4de38370734819a893107ea8864f132ca13d) the references.
