### 2021-04-05 #3141 @IndrajeetPatil
Do you know how can I avoid the duplicated affiliation index (two instances of `1`) here?
