### 2021-04-05 #3141 @xuanxu
> Do you know how can I avoid the duplicated affiliation index (two instances of `1`) here?

The format of the affiliations section in the metadata is incorrect, should be something like:
```yaml
affiliations:
 - name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
   index: 1
```
