+++ paper/paper.bib
@@ -25,20 +25,6 @@
   year = {2001},
 }
 
-@Article{lilienfeld2015,
-  title = {Fifty Psychological and Psychiatric Terms to Avoid: A List of Inaccurate, Misleading, Misused, Ambiguous, and Logically Confused Words and Phrases},
-  volume = {6},
-  issn = {1664-1078},
-  shorttitle = {Fifty Psychological and Psychiatric Terms to Avoid},
-  journal = {Frontiers in Psychology},
-  doi = {10.3389/fpsyg.2015.01100},
-  author = {Scott O. Lilienfeld and Katheryn C. Sauvign{\a'e} and Steven Jay Lynn and Robin L. Cautin and Robert D. Latzman and Irwin D. Waldman},
-  month = {aug},
-  year = {2015},
-  pmid = {26284019},
-  pmcid = {PMC4522609},
-}
-
 @Book{cleveland1985,
   address = {{Monterey, Cal}},
   edition = {1st edition},
