### 2021-04-05 #3141 @IndrajeetPatil
@xuanxu I changed my `.md` file according to your suggestion (https://raw.githubusercontent.com/IndrajeetPatil/ggstatsplot/master/paper/paper.md), but the second `1` has now moved to the end of the affiliation
