### 2021-04-05 #3141 @xuanxu
@IndrajeetPatil You have a dash before the index, so you are creating two affiliation entries.

Current:
```yaml
affiliations:
  - name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
  - index: 1
```

Correct:
```yaml
affiliations:
  - name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
    index: 1
```
