+++ tests/testthat/test-ggcoefstats.R
@@ -166,43 +166,9 @@ test_that(
     pb <- ggplot2::ggplot_build(p)
 
     # tidy dataframe from the function
-    tidy_df <- p$plot_env$tidy_df
+    tidy_df1 <- p$plot_env$tidy_df
 
-    expect_equal(tidy_df$estimate,
-      c(0.8093822, 0.2068347, 0.1176152),
-      tolerance = 1e-3
-    )
-    expect_equal(tidy_df$df[1], 1L)
-    expect_equal(tidy_df$df.error[1], 28L)
-    expect_equal(tidy_df$p.value,
-      c(1.378306e-11, 1.156944e-02, 6.355055e-02),
-      tolerance = 1e-5
-    )
-
-    expect_identical(p$labels$x, "partial eta-squared")
-    expect_identical(p$labels$y, "effect")
-    expect_identical(
-      p$labels$caption,
-      ggplot2::expr(atop(
-        displaystyle(NULL),
-        expr = paste(
-          "AIC = ",
-          "43",
-          ", BIC = ",
-          "50"
-        )
-      ))
-    )
-    expect_null(p$labels$title, NULL)
-    expect_null(p$labels$subtitle, NULL)
-    expect_identical(
-      tidy_df$label,
-      c(
-        "list(~italic(F)(1*\",\"*28)==118.89, ~italic(p)=='1.38e-11', ~widehat(italic(eta)[p]^2)==0.81)",
-        "list(~italic(F)(1*\",\"*28)==7.30, ~italic(p)=='0.012', ~widehat(italic(eta)[p]^2)==0.21)",
-        "list(~italic(F)(1*\",\"*28)==3.73, ~italic(p)=='0.064', ~widehat(italic(eta)[p]^2)==0.12)"
-      )
-    )
+    expect_snapshot(list(tidy_df1, p$labels))
 
     ## partial omega-squared
 
@@ -235,50 +201,19 @@ test_that(
     pb <- ggplot2::ggplot_build(p)
 
     # tidy dataframe from the function
-    tidy_df <- p$plot_env$tidy_df
+    tidy_df2 <- p$plot_env$tidy_df
 
     # tests
-    expect_identical(p$labels$x, "partial omega-squared")
-    expect_identical(p$labels$y, "term")
-    expect_identical(
-      p$labels$caption,
-      ggplot2::expr(atop(
-        displaystyle(paste(italic("Note"), ": From `tidyverse`")),
-        expr = paste("AIC = ", "126", ", BIC = ", "142")
-      ))
-    )
-    expect_identical(p$labels$title, "mammalian sleep")
-    expect_identical(p$labels$subtitle, "Source: `ggplot2` package")
-
-    expect_equal(pb$data[[2]]$x, tidy_df$estimate, tolerance = 0.001)
-    expect_equal(pb$data[[2]]$xmin, tidy_df$conf.low, tolerance = 0.001)
-    expect_equal(pb$data[[2]]$xmax, tidy_df$conf.high, tolerance = 0.001)
+    expect_equal(pb$data[[2]]$x, tidy_df2$estimate, tolerance = 0.001)
+    expect_equal(pb$data[[2]]$xmin, tidy_df2$conf.low, tolerance = 0.001)
+    expect_equal(pb$data[[2]]$xmax, tidy_df2$conf.high, tolerance = 0.001)
     expect_equal(
       pb$data[[2]]$y,
       structure(c(3L, 1L, 2L), class = c("mapped_discrete", "numeric"))
     )
 
-    expect_identical(tidy_df$label, pb$data[[4]]$label)
+    expect_snapshot(list(tidy_df2, p$labels))
 
-    expect_equal(tidy_df$estimate,
-      c(0.30828881, 0.02348073, 0.17365008),
-      tolerance = 0.001
-    )
[…]
