+++ R/attach.R
@@ -1,8 +1,6 @@
 .onAttach <- function(...) {
   packageStartupMessage(
-    'You can cite this package as:
-     Patil, I. (2018). ggstatsplot: "ggplot2" Based Plots with Statistical Details.
-     CRAN. doi:10.5281/zenodo.2074621.
-     Retrieved from https://cran.r-project.org/web/packages/ggstatsplot/index.html'
+    "You can cite this package as:
+     Patil, I. (2021). Visualizations with statistical details: The {'ggstatsplot'} approach. PsyArxiv. doi:10.31234/osf.io/p7mku"
   )
 }
