+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: ggstatsplot
 Title: 'ggplot2' Based Plots with Statistical Details
-Version: 0.7.1.9000
+Version: 0.7.2
 Authors@R: 
     person(given = "Indrajeet",
            family = "Patil",
@@ -29,10 +29,10 @@ Imports:
     ggcorrplot,
     ggplot2,
     ggrepel,
-    ggsignif (>= 0.6.1),
-    insight (>= 0.13.1.1),
-    ipmisc (>= 6.0.0),
-    pairwiseComparisons (>= 3.1.3),
+    ggsignif,
+    insight (>= 0.13.2),
+    ipmisc,
+    pairwiseComparisons,
     paletteer,
     parameters (>= 0.12.0.1),
     patchwork,
