### 2021-04-07 #3141 @csoneson
Thanks @njtierney! We don't impose a fixed due date (and we're a bit more flexible right now given the challenges brought by the pandemic), but we encourage reviewers to post comments as they go through the submission (no need to collect them all in one big review), and to have an initial assessment for the authors to work on within two weeks from the start of the review.
