+++ R/helpers_ggcoefstats_graphics.R
@@ -30,13 +30,13 @@ ggcoefstats_label_maker <- function(tidy_df,
       dplyr::mutate(
         label = dplyr::case_when(
           is.na(df.error) || is.infinite(df.error) ~ paste0(
-            "list(~widehat(italic(beta))=='", format_num(estimate, k),
-            "', ~italic(t)=='", format_num(statistic, k),
+            "list(~widehat(italic(beta))=='", format_value(estimate, k),
+            "', ~italic(t)=='", format_value(statistic, k),
             "', ~italic(p)=='", format_num(p.value, k, TRUE), "')"
           ),
           TRUE ~ paste0(
-            "list(~widehat(italic(beta))=='", format_num(estimate, k),
-            "', ~italic(t)", "('", format_num(df.error, 0L), "')=='", format_num(statistic, k),
+            "list(~widehat(italic(beta))=='", format_value(estimate, k),
+            "', ~italic(t)", "('", format_value(df.error, 0L), "')=='", format_value(statistic, k),
             "', ~italic(p)=='", format_num(p.value, k, TRUE), "')"
           )
         )
@@ -50,8 +50,8 @@ ggcoefstats_label_maker <- function(tidy_df,
     tidy_df %<>%
       dplyr::mutate(
         label = paste0(
-          "list(~widehat(italic(beta))=='", format_num(estimate, k),
-          "', ~italic(z)=='", format_num(statistic, k),
+          "list(~widehat(italic(beta))=='", format_value(estimate, k),
+          "', ~italic(z)=='", format_value(statistic, k),
           "', ~italic(p)=='", format_num(p.value, k, TRUE), "')"
         )
       )
@@ -67,8 +67,8 @@ ggcoefstats_label_maker <- function(tidy_df,
     tidy_df %<>%
       dplyr::mutate(
         label = paste0(
-          "list(~widehat(italic(beta))=='", format_num(estimate, k),
-          "', ~italic(chi)^2~", "('", format_num(df.error, 0L), "')==", format_num(statistic, k),
+          "list(~widehat(italic(beta))=='", format_value(estimate, k),
+          "', ~italic(chi)^2~", "('", format_value(df.error, 0L), "')==", format_value(statistic, k),
           "', ~italic(p)=='", format_num(p.value, k, TRUE), "')"
         )
       )
@@ -85,10 +85,10 @@ ggcoefstats_label_maker <- function(tidy_df,
     tidy_df %<>%
       dplyr::mutate(
         label = paste0(
-          "list(~italic(F)", "('", format_num(df, 0L), "'*\",\"*'", format_num(df.error, 0L),
-          "')=='", format_num(statistic, k),
+          "list(~italic(F)", "('", format_value(df, 0L), "'*\",\"*'", format_value(df.error, 0L),
+          "')=='", format_value(statistic, k),
           "', ~italic(p)=='", format_num(p.value, k, TRUE),
-          "', ~", effsize.text, "=='", format_num(estimate, k), "')"
+          "', ~", effsize.text, "=='", format_value(estimate, k), "')"
         )
       )
   }
