### 2021-04-12 #3167 @whedon
[REVIEW]: Visualizations with statistical details: The 'ggstatsplot' approach
**Submitting author:** @IndrajeetPatil (<a href="http://orcid.org/0000-0003-1995-6531">Indrajeet Patil</a>)
**Repository:** <a href="https://github.com/IndrajeetPatil/ggstatsplot" target ="_blank">https://github.com/IndrajeetPatil/ggstatsplot</a>
**Version:** 0.7.2
