+++ R/ggbarstats.R
@@ -99,40 +99,38 @@ ggbarstats <- function(data,
 
   # if subtitle with results is to be displayed
   if (isTRUE(results.subtitle)) {
-    subtitle_df <-
-      tryCatch(
-        expr = statsExpressions::contingency_table(
-          data = data,
-          x = {{ x }},
-          y = {{ y }},
-          type = type,
-          k = k,
-          paired = paired,
-          ratio = ratio,
-          conf.level = conf.level
-        ),
-        error = function(e) NULL
-      )
+    subtitle_df <- tryCatch(
+      expr = statsExpressions::contingency_table(
+        data = data,
+        x = {{ x }},
+        y = {{ y }},
+        type = type,
+        k = k,
+        paired = paired,
+        ratio = ratio,
+        conf.level = conf.level
+      ),
+      error = function(e) NULL
+    )
 
     if (!is.null(subtitle_df)) subtitle <- subtitle_df$expression[[1]]
 
     # preparing Bayes Factor caption
     if (type != "bayes" && isTRUE(bf.message) && isFALSE(paired)) {
-      caption_df <-
-        tryCatch(
-          expr = statsExpressions::contingency_table(
-            data = data,
-            x = {{ x }},
-            y = {{ y }},
-            type = "bayes",
-            k = k,
-            top.text = caption,
-            sampling.plan = sampling.plan,
-            fixed.margin = fixed.margin,
-            prior.concentration = prior.concentration
-          ),
-          error = function(e) NULL
-        )
+      caption_df <- tryCatch(
+        expr = statsExpressions::contingency_table(
+          data = data,
+          x = {{ x }},
+          y = {{ y }},
+          type = "bayes",
+          k = k,
+          top.text = caption,
+          sampling.plan = sampling.plan,
+          fixed.margin = fixed.margin,
+          prior.concentration = prior.concentration
+        ),
+        error = function(e) NULL
+      )
 
       if (!is.null(caption_df)) caption <- caption_df$expression[[1]]
     }
