+++ R/gghistostats.R
@@ -103,7 +103,7 @@ gghistostats <- function(data,
 
   if (isTRUE(results.subtitle)) {
     # preparing the subtitle with statistical results
-    subtitle_df <-
+    subtitle_df <- tryCatch(
       statsExpressions::one_sample_test(
         data = df,
         x = {{ x }},
@@ -115,13 +115,15 @@ gghistostats <- function(data,
         nboot = nboot,
         tr = tr,
         k = k
-      )
+      ),
+      error = function(e) NULL
+    )
 
-    subtitle <- subtitle_df$expression[[1]]
+    subtitle <- if (!is.null(subtitle_df)) subtitle_df$expression[[1]]
 
     # preparing the BF message
     if (type == "parametric" && isTRUE(bf.message)) {
-      caption_df <-
+      caption_df <- tryCatch(
         statsExpressions::one_sample_test(
           data = df,
           x = {{ x }},
@@ -130,9 +132,11 @@ gghistostats <- function(data,
           bf.prior = bf.prior,
           top.text = caption,
           k = k
-        )
+        ),
+        error = function(e) NULL
+      )
 
-      caption <- caption_df$expression[[1]]
+      caption <- if (!is.null(caption_df)) caption_df$expression[[1]]
     }
   }
 
