+++ DESCRIPTION
@@ -34,7 +34,7 @@ Imports:
     ipmisc,
     pairwiseComparisons,
     paletteer,
-    parameters (>= 0.12.0.1),
+    parameters (>= 0.13.0),
     patchwork,
     performance,
     purrr,
@@ -57,9 +57,6 @@ Suggests:
     tibble,
     vdiffr,
     WRS2
-Remotes:
-    easystats/insight,
-    easystats/parameters
 VignetteBuilder: 
     knitr
 Config/testthat/edition: 3

[…]
