+++ tests/testthat/test-ggbarstats.R
@@ -137,33 +137,6 @@ test_that(
   }
 )
 
-# without enough data ---------------------------------------------------------
-
-test_that(
-  desc = "checking if functions work without enough data",
-  code = {
-    skip_on_cran()
-    set.seed(123)
-
-    # creating a dataframe
-    df <- tibble::tribble(
-      ~x, ~y,
-      "one", "one"
-    )
-
-    # should not work
-    expect_s3_class(
-      suppressWarnings(ggbarstats(
-        data = df,
-        x = x,
-        y = y
-      )),
-      "ggplot"
-    )
-  }
-)
-
-
 # expression output --------------------------------------------------
 
 test_that(
