+++ R/attach.R
@@ -1,6 +1,7 @@
 .onAttach <- function(...) {
   packageStartupMessage(
     "You can cite this package as:
-     Patil, I. (2021). Visualizations with statistical details: The {'ggstatsplot'} approach. PsyArxiv. doi:10.31234/osf.io/p7mku"
+     Patil, I. (2021). Visualizations with statistical details: The 'ggstatsplot' approach.
+     PsyArxiv. doi:10.31234/osf.io/p7mku"
   )
 }
