+++ vignettes/web_only/faq.Rmd
@@ -10,7 +10,6 @@ output:
     toc: true
     warning: FALSE
     message: FALSE
-    number_sections: true
 vignette: >
   %\VignetteIndexEntry{frequently asked questions (FAQ)}
   %\VignetteEngine{knitr::rmarkdown}
@@ -39,18 +38,21 @@ You can cite this package/vignette as:
 citation("ggstatsplot")
 ```
 
+<!-- The sections are numbered manually because `pkgdown` doesn't render -->
+<!-- numbered sections for `rmarkdown::html_vignette` format -->
+
 ---
 
 Following are a few of the common questions asked in GitHub issues and on social
 media platforms.
 
-# I just want the plot, not the statistical details. How can I turn them off?
+# 1. I just want the plot, not the statistical details. How can I turn them off?
 
 All functions in `ggstatsplot` that display results from statistical analysis in
 a subtitle have argument `results.subtitle`. Setting it to `FALSE` will return
 only the plot.
 
-# How can I customize the details contained in the subtitle?
+# 2. How can I customize the details contained in the subtitle?
 
 Sometimes you may not wish include so many details in the subtitle. In that
 case, you can extract the expression and copy-paste only the part you wish to
@@ -76,7 +78,7 @@ ggplot(iris, aes(x = Species, y = Sepal.Length)) +
   )))
 ```
 
-# I am getting `Error in grid.Call` error
+# 3. I am getting `Error in grid.Call` error
 
 Sometimes, if you are working in `RStudio`, you might see the following error-
 
@@ -91,7 +93,7 @@ those solutions solve your problem.
 If not, pray to the old and the new gods, and try again. It might just solve
 your problem. `r emo::ji("shrug")`
 
-# Why do I get only plot in return but not the subtitle/caption?
+# 4. Why do I get only plot in return but not the subtitle/caption?
 
 In order to prevent the function from failing when statistical analysis fails,
 functions in `ggstatsplot` default to first attempting to run the analysis and
@@ -119,7 +121,7 @@ df <- data.frame(x = c("a", "b"), y = c(1, 2))
 two_sample_test(data = df, x = x, y = y)
 ```
 
-# What statistical test was carried out?
+# 5. What statistical test was carried out?
 
 In case you are not sure what was the statistical test that produced the results
 shown in the subtitle of the plot, the best way to get that information is to
@@ -129,7 +131,7 @@ associated vignette.
 Summary of all analysis is handily available in `README`: 
 <https://github.com/IndrajeetPatil/ggstatsplot/blob/master/README.md>
 
-# How can I use `ggstatsplot` functions in a `for` loop?
+# 6. How can I use `ggstatsplot` functions in a `for` loop?
 
 Given that all functions in `ggstatsplot` use tidy evaluation, running these
 functions in a `for` loop requires minor adjustment to how inputs are entered:
@@ -156,7 +158,7 @@ dataframe in what you want to do, I will recommend you to have a look at
 
 <https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/purrr_examples.html#repeating-function-execution-across-multiple-columns-in-a-dataframe-1>
 
-# How can I have uniform Y-axes ranges in `grouped_` functions?
+# 7. How can I have uniform Y-axes ranges in `grouped_` functions?
 
 ```{r grouped_y_axes, fig.height=10, fig.width=8}
 # setup
@@ -190,7 +192,7 @@ combine_plots(
 )
 ```
 
-# Does `ggstatsplot` work with `plotly`?
+# 8. Does `ggstatsplot` work with `plotly`?
 
[…]
