### 2021-04-12 76c332c
remove lifecycle badges; cran resubmission
