+++ R/grouped_ggcorrmat.R
@@ -4,7 +4,7 @@
 #'
 #' @description
 #'
-#' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
+#'
 #'
 #' Helper function for `ggstatsplot::ggcorrmat` to apply this function across
 #' multiple levels of a given factor and combining the resulting plots using
