+++ R/grouped_ggscatterstats.R
@@ -4,7 +4,7 @@
 #'
 #' @description
 #'
-#' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
+#'
 #'
 #' Grouped scatterplots from `ggplot2` combined with marginal distribution plots
 #' with statistical details added as a subtitle.
