+++ R/ggdotplotstats.R
@@ -3,7 +3,7 @@
 #'
 #' @description
 #'
-#' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
+#'
 #'
 #' A dot chart (as described by William S. Cleveland) with statistical details
 #' from one-sample test included in the plot as a subtitle.
