+++ R/ggscatterstats.R
@@ -3,7 +3,7 @@
 #'
 #' @description
 #'
-#' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
+#'
 #'
 #' Scatterplots from `ggplot2` combined with marginal
 #' histograms/boxplots/density plots with statistical details added as a
