+++ R/ggcoefstats.R
@@ -3,7 +3,7 @@
 #'
 #' @description
 #'
-#' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
+#'
 #'
 #' Plot with the regression coefficients' point estimates as dots with
 #' confidence interval whiskers and other statistical details included as
