+++ R/ggpiestats.R
@@ -3,7 +3,7 @@
 #'
 #' @description
 #'
-#' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
+#'
 #'
 #' Pie charts for categorical data with statistical details included in the plot
 #' as a subtitle.
