+++ README.md
@@ -225,35 +225,9 @@ comparisons <br> ✅ Bayesian hypothesis-testing <br> ✅ Bayesian
 estimation <br>
 
 A number of other arguments can be specified to make this plot even more
-informative or change some of the default options.
-
-``` r
-# for reproducibility
-set.seed(123)
-library(ggplot2)
-
-# plot
-ggbetweenstats(
-  data = ToothGrowth,
-  x = supp,
-  y = len,
-  type = "r", # robust statistics
-  k = 3, # number of decimal places for statistical results
-  xlab = "Supplement type", # label for the x-axis variable
-  ylab = "Tooth length", # label for the y-axis variable
-  title = "The Effect of Vitamin C on Tooth Growth", # title text for the plot
-  ggtheme = ggthemes::theme_fivethirtyeight(), # choosing a different theme
-  ggstatsplot.layer = FALSE, # turn off `ggstatsplot` theme layer
-  package = "wesanderson", # package from which color palette is to be taken
-  palette = "Darjeeling1" # choosing a different color palette
-)
-```
-
-<img src="man/figures/README-ggbetweenstats2-1.png" width="100%" />
-
-Additionally, there is also a `grouped_` variant of this function that
-makes it easy to repeat the same operation across a **single** grouping
-variable:
+informative or change some of the default options. Additionally, there
+is also a `grouped_` variant of this function that makes it easy to
+repeat the same operation across a **single** grouping variable:
 
 ``` r
 # for reproducibility
@@ -261,10 +235,7 @@ set.seed(123)
 
 # plot
 grouped_ggbetweenstats(
-  data = dplyr::filter(
-    .data = movies_long,
-    genre %in% c("Action", "Action Comedy", "Action Drama", "Comedy")
-  ),
+  data = dplyr::filter(movies_long, genre %in% c("Action", "Comedy")),
   x = mpaa,
   y = length,
   grouping.var = genre, # grouping variable
@@ -278,12 +249,12 @@ grouped_ggbetweenstats(
   caption = substitute(paste(italic("Source"), ": IMDb (Internet Movie Database)")),
   palette = "default_jama",
   package = "ggsci",
-  plotgrid.args = list(nrow = 2),
+  plotgrid.args = list(nrow = 1),
   annotation.args = list(title = "Differences in movie length by mpaa ratings for different genres")
 )
 ```
 
-<img src="man/figures/README-ggbetweenstats3-1.png" width="100%" />
+<img src="man/figures/README-ggbetweenstats2-1.png" width="100%" />
 
 Note here that the function can be used to tag outliers!
 
@@ -510,10 +481,7 @@ set.seed(123)
 
 # plot
 grouped_gghistostats(
-  data = dplyr::filter(
-    .data = movies_long,
-    genre %in% c("Action", "Action Comedy", "Action Drama", "Comedy")
-  ),
+  data = dplyr::filter(movies_long, genre %in% c("Action", "Comedy")),
   x = budget,
   test.value = 50,
   type = "nonparametric",
@@ -524,7 +492,7 @@ grouped_gghistostats(
   ggtheme = ggthemes::theme_tufte(),
   # modify the defaults from `ggstatsplot` for each plot
   ggplot.component = ggplot2::labs(caption = "Source: IMDB.com"),
-  plotgrid.args = list(nrow = 2),
+  plotgrid.args = list(nrow = 1),
   annotation.args = list(title = "Movies budgets for different genres")
 )
 ```
@@ -670,7 +638,7 @@ set.seed(123)
 
 # plot
 ggscatterstats(
-  data = dplyr::filter(.data = movies_long, genre == "Action"),
+  data = dplyr::filter(movies_long, genre == "Action"),
   x = budget,
[…]
