+++ README.Rmd
@@ -18,7 +18,7 @@ options(
 
 knitr::opts_chunk$set(
   collapse = TRUE,
-  dpi = 300,
+  dpi = 150, # change to 300 once on CRAN
   warning = FALSE,
   message = FALSE,
   out.width = "100%",
@@ -232,43 +232,17 @@ ggbetweenstats(
 `r emo::ji("check")` Bayesian estimation <br>
 
 A number of other arguments can be specified to make this plot even more
-informative or change some of the default options. 
+informative or change some of the default options. Additionally, there is also a
+`grouped_` variant of this function that makes it easy to repeat the same
+operation across a **single** grouping variable:
 
-```{r ggbetweenstats2}
-# for reproducibility
-set.seed(123)
-library(ggplot2)
-
-# plot
-ggbetweenstats(
-  data = ToothGrowth,
-  x = supp,
-  y = len,
-  type = "r", # robust statistics
-  k = 3, # number of decimal places for statistical results
-  xlab = "Supplement type", # label for the x-axis variable
-  ylab = "Tooth length", # label for the y-axis variable
-  title = "The Effect of Vitamin C on Tooth Growth", # title text for the plot
-  ggtheme = ggthemes::theme_fivethirtyeight(), # choosing a different theme
-  ggstatsplot.layer = FALSE, # turn off `ggstatsplot` theme layer
-  package = "wesanderson", # package from which color palette is to be taken
-  palette = "Darjeeling1" # choosing a different color palette
-)
-```
-
-Additionally, there is also a `grouped_` variant of this function that makes it
-easy to repeat the same operation across a **single** grouping variable:
-
-```{r ggbetweenstats3, fig.height = 14, fig.width = 12}
+```{r ggbetweenstats2, fig.height = 8, fig.width = 12}
 # for reproducibility
 set.seed(123)
 
 # plot
 grouped_ggbetweenstats(
-  data = dplyr::filter(
-    .data = movies_long,
-    genre %in% c("Action", "Action Comedy", "Action Drama", "Comedy")
-  ),
+  data = dplyr::filter(movies_long, genre %in% c("Action", "Comedy")),
   x = mpaa,
   y = length,
   grouping.var = genre, # grouping variable
@@ -282,7 +256,7 @@ grouped_ggbetweenstats(
   caption = substitute(paste(italic("Source"), ": IMDb (Internet Movie Database)")),
   palette = "default_jama",
   package = "ggsci",
-  plotgrid.args = list(nrow = 2),
+  plotgrid.args = list(nrow = 1),
   annotation.args = list(title = "Differences in movie length by mpaa ratings for different genres")
 )
 ```
@@ -503,16 +477,13 @@ gghistostats(
 There is also a `grouped_` variant of this function that makes it
 easy to repeat the same operation across a **single** grouping variable:
 
-```{r gghistostats2, fig.height = 10, fig.width = 10}
+```{r gghistostats2, fig.height = 6, fig.width = 10}
 # for reproducibility
 set.seed(123)
 
 # plot
 grouped_gghistostats(
-  data = dplyr::filter(
-    .data = movies_long,
-    genre %in% c("Action", "Action Comedy", "Action Drama", "Comedy")
-  ),
+  data = dplyr::filter(movies_long, genre %in% c("Action", "Comedy")),
   x = budget,
   test.value = 50,
   type = "nonparametric",
@@ -523,7 +494,7 @@ grouped_gghistostats(
   ggtheme = ggthemes::theme_tufte(),
   # modify the defaults from `ggstatsplot` for each plot
   ggplot.component = ggplot2::labs(caption = "Source: IMDB.com"),
-  plotgrid.args = list(nrow = 2),
+  plotgrid.args = list(nrow = 1),
[…]
