### 2021-04-13 51956f5
back to 300 dpi; simplify FAQ item
