### 2021-04-14 aea2a0d
bump to devel; minor restructuring of vignettes
