+++ vignettes/web_only/ggdotplotstats.Rmd
@@ -49,7 +49,7 @@ this vignette we will explore several examples of how to use it.
 This function is a sister function of `gghistostats` with the difference being
 it expects a labeled numeric variable.
 
-## Distribution of a sample with `ggdotplotstats`
+# Distribution of a sample with `ggdotplotstats`
 
 Let's begin with a very simple example from the `ggplot2` package
 (`ggplot2::mpg`), a subset of the fuel economy data that the EPA makes available
@@ -96,7 +96,7 @@ ggstatsplot::ggdotplotstats(
 )
 ```
 
-## Grouped analysis with `grouped_ggdotplotstats`
+# Grouped analysis with `grouped_ggdotplotstats`
 
 What if we want to do the same analysis separately for different engines with
 different numbers of cylinders?
@@ -129,13 +129,13 @@ ggstatsplot::grouped_ggdotplotstats(
   type = "bayes", # Bayesian test
   test.value = 15.5,
   ggtheme = ggthemes::theme_fivethirtyeight(),
-  # arguments relevant for ggstatsplot::combine_plots
+  # arguments relevant for `ggstatsplot::combine_plots`
   annotation.args = list(title = "Fuel economy data"),
   plotgrid.args = list(nrow = 2)
 )
 ```
 
-## Grouped analysis with `purrr`
+# Grouped analysis with `purrr`
 
 Although this is a quick and dirty way to explore a large amount of data with
 minimal effort, it does come with an important limitation: reduced flexibility.
@@ -148,7 +148,7 @@ use `purrr`.
 See the associated vignette here:
 <https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/purrr_examples.html>
 
-## Summary of tests
+# Summary of tests
 
 The central tendency measure displayed will depend on the statistics:
 
@@ -180,7 +180,7 @@ Non-parametric | *r* (rank-biserial correlation) | <font color="green">Yes<font>
 Robust | trimmed mean | <font color="green">Yes<font> | `trimcibt` (custom)
 Bayes Factor | $\delta_{posterior}$ | <font color="green">Yes<font> | `bayestestR::describe_posterior`
 
-## Effect size interpretation
+# Effect size interpretation
 
 To see how the effect sizes displayed in these tests can be interpreted, see:
 <https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/effsize_interpretation.html>
