+++ NEWS.md
@@ -1,3 +1,5 @@
+# ggstatsplot 0.7.2.9000
+
 # ggstatsplot 0.7.2
 
 MAJOR CHANGES
