+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: ggstatsplot
 Title: 'ggplot2' Based Plots with Statistical Details
-Version: 0.7.2
+Version: 0.7.2.9000
 Authors@R: 
     person(given = "Indrajeet",
            family = "Patil",

[…]
