### 2021-04-15 894cd90
add lifecycle badges for functions in README and vignettes
