+++ vignettes/web_only/ggcorrmat.Rmd
@@ -41,7 +41,9 @@ citation("ggstatsplot")
 
 ---
 
-The function `ggstatsplot::ggcorrmat` provides a quick way to produce
+Lifecycle: [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://lifecycle.r-lib.org/articles/stages.html)
+
+The function `ggcorrmat` provides a quick way to produce
 **publication-ready correlation matrix** (aka *correlalogram*) plot. The function
 can also be used for quick **data exploration**. In addition to the plot, it can
 also be used to get a correlation coefficient matrix or the associated *p*-value
