+++ vignettes/web_only/ggscatterstats.Rmd
@@ -36,11 +36,13 @@ knitr::opts_chunk$set(
 )
 ```
 
-The function `ggstatsplot::ggscatterstats` is meant to provide a
-**publication-ready scatterplot** with all statistical details included in the
-plot itself to show association between two continuous variables. This function
-is also helpful during the **data exploration** phase. We will see examples of how
-to use this function in this vignette with the `ggplot2movies` dataset.
+Lifecycle: [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://lifecycle.r-lib.org/articles/stages.html)
+
+The function `ggscatterstats` is meant to provide a **publication-ready
+scatterplot** with all statistical details included in the plot itself to show
+association between two continuous variables. This function is also helpful
+during the **data exploration** phase. We will see examples of how to use this
+function in this vignette with the `ggplot2movies` dataset.
 
 To begin with, here are some instances where you would want to use
 `ggscatterstats`-
