### 2021-04-16 a9312e4
add reporting to rest of the vignettes
