### 2021-04-21 IndrajeetPatil/ggstatsplot#563 @kevinrue
JOSS review: typo in paper
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L44

"to evaluate [the] validity of a model"

For https://github.com/openjournals/joss-reviews/issues/3167
