### 2021-04-21 IndrajeetPatil/ggstatsplot#564 @kevinrue
JOSS review: odd phrasing
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L44

"if the said analysis"

there is no mention of "analysis" so far

"if the analysis" or "if an analysis" might sound better here

For https://github.com/openjournals/joss-reviews/issues/3167
