### 2021-04-21 IndrajeetPatil/ggstatsplot#565 @kevinrue
JOSS review: suggestion of edit for paper
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L49

"create densely informative `ggplot2`-based visualizations"

almost sounds neative, as in "overly informative"

I'd suggest

"enriches `ggplot2`-based visualizations with the results from
statistical analysis"

For https://github.com/openjournals/joss-reviews/issues/3167
