### 2021-04-21 IndrajeetPatil/ggstatsplot#566 @kevinrue
JOSS review: paper edit suggestion
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L57

"and modeling in its turn can suggest a different visualization method"

"and in turn, modelling can suggest a different visualization method"

For https://github.com/openjournals/joss-reviews/issues/3167
