### 2021-04-21 IndrajeetPatil/ggstatsplot#567 @kevinrue
JOSS review: Consider removing bold font
Traditionally schorlarly articles seldom use bold font.

I would remove the bold font used in those places, where the words themselves already put sufficient emphasis on the message, in my opinion:

- https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L51
- https://github.com/IndrajeetPatil/ggstatsplot/blob/master/paper/paper.Rmd#L73
- https://github.com/IndrajeetPatil/ggstatsplot/blob/master/paper/paper.Rmd#L78

For https://github.com/openjournals/joss-reviews/issues/3167
