### 2021-04-21 IndrajeetPatil/ggstatsplot#568 @kevinrue
JOSS review: paper edit suggestion
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L83

> removes the trade-off often faced by researchers between ease versus
> thoroughness of exploring data and further cements good data

I would suggest

> removes the trade-off often faced by researchers between ease [and]
> thoroughness of [data exploration] and further cements good data

For https://github.com/openjournals/joss-reviews/issues/3167
