### 2021-04-21 IndrajeetPatil/ggstatsplot#569 @kevinrue
JOSS review: paper edit suggestion
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L85

> sanitation/exploration habits.

I would suggest removing "sanitation/", i.e.

> exploration habits.

For https://github.com/openjournals/joss-reviews/issues/3167
