### 2021-04-21 IndrajeetPatil/ggstatsplot#570 @kevinrue
JOSS review: suggested paper edit
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L108

> e. encourages researchers/readers to evaluate statistical assumptions of a model
   in the context of the underlying data (Figure 2),

I suggest:

> e. encourages researchers [and] readers to evaluate statistical assumptions of a model
   in the context of the underlying data (Figure 2),

For https://github.com/openjournals/joss-reviews/issues/3167
