### 2021-04-21 IndrajeetPatil/ggstatsplot#571 @kevinrue
JOSS review: paper edit suggestion
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L111

> f. is easy and simple enough that someone with little-to-no coding experience
   can use it without making an error and may even encourage beginners to
   programmatically (instead of using GUI software) analyze data.

I suggest

> f. is easy and simple enough that someone with little-to-no coding experience
   can use it without making an error and may even encourage beginners to
   programmatically [analyze data, instead of using GUI software].

For https://github.com/openjournals/joss-reviews/issues/3167
