### 2021-04-21 IndrajeetPatil/ggstatsplot#573 @kevinrue
JOSS review: paper edit suggestion
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/paper/paper.Rmd#L133

Currently:

> In the spirit of honest and open science, requests/tips for fixes, feature
updates, as well as general questions and concerns via direct interaction with
contributors and developers are encouraged by filing an
[issue](https://github.com/IndrajeetPatil/ggstatsplot/issues) while respecting
[*Contribution Guidelines*](https://indrajeetpatil.github.io/ggstatsplot/CONTRIBUTING.html).

Suggested edits:

> In the spirit of honest and open science, requests and suggestions for fixes, feature
updates, as well as general questions and concerns are encouraged via direct interaction with
contributors and developers by filing an
[issue](https://github.com/IndrajeetPatil/ggstatsplot/issues) while respecting
[*Contribution Guidelines*](https://indrajeetpatil.github.io/ggstatsplot/CONTRIBUTING.html).

For https://github.com/openjournals/joss-reviews/issues/3167
