### 2021-04-21 IndrajeetPatil/ggstatsplot#574 @kevinrue
JOSS review: author list
To date, the statistics of repository contributors report that @ibecav has made 200 commits  to the repository adding 9,644 lines and removing 6,479 lines.

See:

- https://github.com/IndrajeetPatil/ggstatsplot/graphs/contributors
- https://github.com/IndrajeetPatil/ggstatsplot/commits?author=ibecav

This seems worthy of adding the contributor at least as a package contributor in the DESCRIPTION file (ctb, see https://cran.r-project.org/doc/manuals/R-exts.html#The-DESCRIPTION-file), unless they deserve authorship (aut), in which case that should also be reflected as authorship in the published article.

For https://github.com/openjournals/joss-reviews/issues/3167
