+++ paper/paper.Rmd
@@ -41,21 +41,21 @@ knitr::opts_chunk$set(
 
 Graphical displays can reveal problems in a statistical model that might not be
 apparent from purely numerical summaries. Such visualizations can also be
-helpful for the reader to evaluate validity of a model if the said analysis is
-reported in a scholarly publication/report. But, given the onerous costs
-involved, researchers can avoid preparing information-rich graphics and
-exploring several statistical approaches/tests available. The `ggstatsplot`
-package in R programming language [@base2021] provides a one-line syntax to
-create densely informative `ggplot2`-based visualizations with the results from
-statistical analysis embedded in the visualization itself. In doing so, the
-package helps researchers adopt a **rigorous, reliable, and robust** data
-exploratory and reporting workflow.
+helpful for the reader to evaluate the validity of a model if it is reported in
+a scholarly publication/report. But, given the onerous costs involved,
+researchers can avoid preparing information-rich graphics and exploring several
+statistical approaches/tests available. The `ggstatsplot` package in R
+programming language [@base2021] provides a one-line syntax to enrich
+`ggplot2`-based visualizations with the results from statistical analysis
+embedded in the visualization itself. In doing so, the package helps researchers
+adopt a rigorous, reliable, and robust data exploratory and reporting
+workflow.
 
 # Statement of Need
 
 In a typical data analysis workflow, data visualization and statistical modeling
-are two different phases: visualization informs modeling, and modeling in its
-turn can suggest a different visualization method, and so on and so forth
+are two different phases: visualization informs modeling, and in turn, modeling
+can suggest a different visualization method, and so on and so forth
 [@wickham2016r]. The central idea of `ggstatsplot` is simple: combine these two
 phases into one in the form of an informative graphic with statistical details.
 
@@ -70,19 +70,18 @@ library(ggstatsplot)
 ggbetweenstats(penguins, species, body_mass_g)
 ```
 
-As can be seen, with a **single** line of code, the function produces details
-about descriptive statistics, inferential statistics, effect size estimate and
-its uncertainty, pairwise comparisons, Bayesian hypothesis testing, Bayesian
+As can be seen, with a single line of code, the function produces details about
+descriptive statistics, inferential statistics, effect size estimate and its
+uncertainty, pairwise comparisons, Bayesian hypothesis testing, Bayesian
 posterior estimate and its uncertainty. Moreover, these details are juxtaposed
 with informative and well-labeled visualizations. The defaults are designed to
-follow best practices in **both** data visualization [@cleveland1985;
+follow best practices in both data visualization [@cleveland1985;
 @grant2018data; @healy2018data; @tufte2001; @wilke2019fundamentals] and
 (Frequentist/Bayesian) statistical reporting [@apa2019; @van2020jasp]. Without
 `ggstatsplot`, getting these statistical details and customizing a plot would
 require significant amount of time and effort In other words, this package
-removes the trade-off often faced by researchers between ease versus
-thoroughness of exploring data and further cements good data
-sanitation/exploration habits.
+removes the trade-off often faced by researchers between ease and thoroughness
+of data exploration and further cements good data exploration habits.
 
 Internally, data cleaning is carried out using `tidyverse` [@Wickham2019], while
 statistical analysis is carried out via `statsExpressions` [@Patil2021] and
@@ -105,14 +104,14 @@ c. highlights the importance of the effect by providing effect size measures by
 
 d. provides an easy way to evaluate *absence* of an effect using Bayes factors,
 
-e. encourages researchers/readers to evaluate statistical assumptions of a model
-   in the context of the underlying data (Figure 2),
+e. encourages researchers and readers to evaluate statistical assumptions of a
[…]
