### 2021-04-21 IndrajeetPatil/ggstatsplot#575 @kevinrue
JOSS review: clarity of supported features table in README
https://github.com/IndrajeetPatil/ggstatsplot/blob/22e7a844e90acc4632f74b0361c68ddfcfd55701/README.Rmd#L159

Currently, there is little contrast between "Yes" and `NA` in this table.

I suggest using icons, as it is done further below (e.g. section "Defaults return").

For instance, you could:

- replace "Yes" by :white_check_mark:
- replace "NA" by :x:

Which would instantly visually distinguish features that are supported from those that aren't yet.

For https://github.com/openjournals/joss-reviews/issues/3167
