+++ README.Rmd
@@ -156,25 +156,25 @@ supported in this package-
 
 Functions | Description | Parametric | Non-parametric | Robust | Bayesian
 ------- | ------------------ | ---- | ----- | ----| ----- 
-`ggbetweenstats` | Between group/condition comparisons | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font>
-`ggwithinstats` | Within group/condition comparisons | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font>
-`gghistostats`, `ggdotplotstats` | Distribution of a numeric variable | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font>
-`ggcorrmat` | Correlation matrix | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font>
-`ggscatterstats` | Correlation between two variables | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font>
-`ggpiestats`, `ggbarstats` | Association between categorical variables | <font color="green">Yes</font> | `NA` | `NA` | <font color="green">Yes</font>
-`ggpiestats`, `ggbarstats` | Equal proportions for categorical variable levels | <font color="green">Yes</font> | `NA` | `NA` | <font color="green">Yes</font>
-`ggcoefstats` | Regression model coefficients | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font> | <font color="green">Yes</font>
-`ggcoefstats` | Random-effects meta-analysis | <font color="green">Yes</font> | `NA` | <font color="green">Yes</font> | <font color="green">Yes</font>
+`ggbetweenstats` | Between group/condition comparisons | ✅ | ✅ | ✅ | ✅
+`ggwithinstats` | Within group/condition comparisons | ✅ | ✅ | ✅ | ✅
+`gghistostats`, `ggdotplotstats` | Distribution of a numeric variable | ✅ | ✅ | ✅ | ✅
+`ggcorrmat` | Correlation matrix | ✅ | ✅ | ✅ | ✅
+`ggscatterstats` | Correlation between two variables | ✅ | ✅ | ✅ | ✅
+`ggpiestats`, `ggbarstats` | Association between categorical variables | ✅ | ✅ | ❌| ✅
+`ggpiestats`, `ggbarstats` | Equal proportions for categorical variable levels | ✅ | ❌| ❌| ✅
+`ggcoefstats` | Regression model coefficients | ✅ | ✅ | ✅ | ✅
+`ggcoefstats` | Random-effects meta-analysis | ✅ | ❌| ✅ | ✅
 
 Summary of Bayesian analysis
 
 Analysis | Hypothesis testing | Estimation
 ------------------ | ---------- | ---------
-(one/two-sample) t-test | <font color="green">Yes</font> | <font color="green">Yes</font>
-one-way ANOVA | <font color="green">Yes</font> |<font color="green">Yes</font>
-correlation | <font color="green">Yes</font> | <font color="green">Yes</font>
-(one/two-way) contingency table | <font color="green">Yes</font> | <font color="green">Yes</font>
-random-effects meta-analysis | <font color="green">Yes</font> | <font color="green">Yes</font>
[…]
