### 2021-04-21 IndrajeetPatil/ggstatsplot#576 @kevinrue
Hm.. thinking about this more, `ggthemes` is only used in the user session, i.e. it is not a dependency of `ggstatsplot`.
So it is natural to leave it to users to install, if they wish to use it.
Closing the issue
