### 2021-04-21 IndrajeetPatil/ggstatsplot#576 @IndrajeetPatil
This is by design.

> Error in library(afex) : there is no package called ‘afex’

As you mention, it is in `Suggests` and the package will suggest the user to install it if the function can't find it:

https://github.com/IndrajeetPatil/statsExpressions/blob/0360a69d77f426832e44e73e2dfa889f40a0a41b/R/oneway_anova.R#L198

> even the full installation command in the `README` does not install `ggthemes`

This is because `ggthemes` plays no role at all in `ggstatsplot` functions. This example is just to show how one can aesthetically modify the default if one wishes to. All `ggstatsplot` functions have `ggtheme` argument, whose function is to allow users to install whatever theme package they wish to install. So if someone is _choosing_ to use a theme from `ggthemes` package, it is safe to assume that they have it installed.
