### 2021-04-21 IndrajeetPatil/ggstatsplot#577 @kevinrue
I solved it by installing the font from https://fonts.google.com/specimen/Titillium+Web
