### 2021-04-21 IndrajeetPatil/ggstatsplot#577 @kevinrue
I'll let you decide whether this should be documented anywhere, or whether it's OK to close the issue as it is.
