### 2021-04-21 IndrajeetPatil/ggstatsplot#577 @IndrajeetPatil
FAQ 3: https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/faq.html#3-i-am-getting-error-in-grid-call-error

This is one of the most common error encountered while using `ggplot2` on `MacOS` and has nothing to do with `ggstatsplot` _per se_.

For example,
- https://stackoverflow.com/questions/10581440/error-in-grid-calll-textbounds-as-graphicsannotxlabel-xx-xy-polygon
- https://community.rstudio.com/t/comeback-of-the-infamous-error-in-grid-call-c-textbounds-as-graphicsannot-x-label-x-x-x-y-polygon-edge-not-found-bug/54555
- https://community.rstudio.com/t/still-fighting-with-grid-call-c-textbounds-as-graphicsannot-x-label-x-x-x-y-polygon-edge-not-found-error/65559
- https://github.com/tidyverse/ggplot2/issues/2252
and the list goes on.

Not always clear where it is coming from, so I have no instructions about this error.
