### 2021-04-21 IndrajeetPatil/ggstatsplot#574 @IndrajeetPatil
Yes, it is true that @ibecav contributed during the initial phase to `ggstatsplot`'s development, but things have changed dramatically since then. He was indeed listed as a co-author initially but was later removed (and is not be listed as a co-author on this paper and package) for the following reasons:

1. `ggstatsplot`-universe

Although `ggstatsplot` used to be a single package when he contributed (March-June 2019), it was then split into multiple smaller R packages (`ipmisc`, `pairwiseComparisons`, `statsExpressions`, `broomExtra`, and `tidyBF`) for development and maintenance convenience. As a result, many of his contributions were dispersed across these packages, and were further outright removed/replaced during major code refactoring (see Point 3 below) that took place since he stopped contributing (June 2019).

2. Size of contributions

You mention that he has added 9,644 lines and removing 6,479 lines.

Across the 6 packages (`ggstatsplot`, `ipmisc`, `pairwiseComparisons`, `statsExpressions`, `broomExtra`, and `tidyBF`) that supported or supports `ggstatsplot` package, the tally of total contributions (mostly from me) to GitHub repositories have been following:
(on 21.04.2021 at 15:49 Berlin time)

Package | Additions | Deletions
----------- | ------------ | ------------
`ggstatsplot` | 790,581 ++ | 654,965 --
`statsExpressions` | 143,192 ++  | 127,042 --
`pairwiseComparisons` | 48,260 ++ | 41,735 --
`ipmisc` |  28,056 ++ | 24,502 --
`tidyBF` | 29,874 ++ | 23,316 --
`broomExtra` | 61,972 ++ | 54,657 --
**Total** |  **1,101,935 ++** | **926,217 --**

So, currently, his contributions count for **0.87%** (9,644 / 1,101,935) and **0.70%** (6,479 / 926,217) of all codebase that supported or supports `ggstatsplot` package. This is not even counting that this is an ongoing project and these contributions would continue to dilute further.

Therefore, I do not consider these contributions significant enough for co-authorship on the paper.

3. Refactoring

Since he stopped contributing in June 2019, the entire codebase of `ggstatsplot`, which was over 12,000 lines of code at the time was slowly, painfully, and painstakingly refactored to the current codebase of only about 2,000 lines of code. This was a major refactoring that removed/replaced most of his contributions (e.g., a function to compute confidence intervals for Cramer's _V_, using `ez` and `MBESS` for ANOVA, etc.) with other packages/ functions (`effectsize::cramers_v`, `afex::ez_aov` and `effectsize::eta_squared`, etc.). Some other functionality he had added to this package (`ggbarstats`) is now also re-implemented in functions from his own personal package (https://cran.r-project.org/web/packages/CGPfunctions/vignettes/Using-PlotXTabs2.html).
