### 2021-04-21 IndrajeetPatil/ggstatsplot#577 @kevinrue
Noted - hopefully this threads and the links that you provided will give other users good places to start looking for answers.

I came back to this thread because I just ran into a similar error again, because of another font missing on my computer

```r
ggscatterstats(
  data = dplyr::filter(movies_long, genre == "Action"),
  x = budget,
  y = rating,
  type = "robust", # type of test that needs to be run
  xlab = "Movie budget (in million/ US$)", # label for x axis
  ylab = "IMDB rating", # label for y axis
  label.var = title, # variable for labeling data points
  label.expression = rating < 5 & budget > 100, # expression that decides which points to label
  title = "Movie budget and IMDB rating (action)", # title text for the plot
  caption = expression(paste(italic("Note"), ": IMDB stands for Internet Movie DataBase")),
  ggtheme = hrbrthemes::theme_ipsum_ps(), # choosing a different theme
  ggstatsplot.layer = FALSE, # turn off `ggstatsplot` theme layer
  marginal.type = "boxplot", # type of marginal distribution to be displayed
  xfill = "pink", # color fill for x-axis marginal distribution
  yfill = "#009E73" # color fill for y-axis marginal distribution
)
```

```
  no font could be found for family "IBMPlexSans"
```

Also, speaking about RStudio, from past experience, I noticed that resizing the "Plots" pane often helps, especially increasing the size of the pane, so that RStudio gets as much space as possible to fits margins.
