### 2021-04-21 IndrajeetPatil/ggstatsplot#577 @kevinrue
Also: users should make sure that they restart RStudio after installing new fonts.
