### 2021-04-21 IndrajeetPatil/ggstatsplot#579 @kevinrue
JOSS review: increase size of 'X' in ggcorrmat?
More of a question than a suggestion:

How about increasing the 'X' that marks non-significant correlations?

Right now, it's actually a bit difficult to spot the X's because they're about the same size at the correlation value.

I'm wondering whether making the 'X' bigger would achieve two benefits:

- easier to spot the X's
- easier to read the non-significant correlation values (for what they're worth ...) that are hidden behind those X's

****
 For https://github.com/openjournals/joss-reviews/issues/3167
