### 2021-04-21 IndrajeetPatil/ggstatsplot#577 @IndrajeetPatil
> Also, speaking about RStudio, from past experience, I noticed that resizing the "Plots" pane often helps, especially increasing the size of the pane, so that RStudio gets as much space as possible to fits margins.

That's a helpful tip, thanks!
