### 2021-04-21 IndrajeetPatil/ggstatsplot#579 @IndrajeetPatil
Thanks for the suggestion. This is a good point. Here is the changed behavior:

``` r
# for reproducibility
set.seed(123)
library(ggstatsplot)

ggcorrmat(
  data = ggplot2::msleep,
  colors = c("#B2182B", "white", "#4D4D4D"),
  title = "Correlalogram for mammals sleep dataset",
  subtitle = "sleep units: hours; weight units: kilograms"
)
```

![](https://i.imgur.com/O8VYs58.png)
