### 2021-04-21 IndrajeetPatil/ggstatsplot#574 @kevinrue
Thanks for the thorough historical view of the package evolution.

I thought as well that authorship was probably too much. I mostly wondered leaned toward the status of contributor (ctb), which would not affect authorship in the paper itself.

Having said that, I agree that the overall proportion of contribution, and the fact that he is ceased to contribute to the software development effort a while ago, is enough justification for me.

Thanks again for taking the time to explain.
