+++ tests/testthat/test-ggcorrmat.R
@@ -33,44 +33,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-
-    # checking layers
-    expect_equal(pb$plot$layers[[3]]$aes_params$shape, "cross")
-    expect_equal(pb$plot$layers[[3]]$aes_params$size, 14L)
-    expect_identical(pb$plot$layers[[3]]$aes_params$colour, "white")
-
-    # checking plot labels
-    expect_identical(p$labels$title, "Iris dataset")
-    expect_identical(p$labels$subtitle, "By Edgar Anderson")
-    expect_identical(
-      p_legend_title,
-      ggplot2::expr(atop(
-        atop(scriptstyle(bold("sample sizes:")), italic(n) ~
-        "=" ~ "150"),
-        atop(
-          scriptstyle(bold("correlation:")),
-          "Pearson"
-        )
-      ))
-    )
-    expect_identical(
-      p$labels$caption,
-      ggplot2::expr(atop(
-        displaystyle(NULL),
-        expr = paste(
-          bold("X"),
-          " = non-significant at ",
-          italic("p"),
-          " < ",
-          0.001,
-          " (Adjustment: ",
-          "FDR",
-          ")"
-        )
-      ))
-    )
-    expect_null(p$labels$xlab, NULL)
-    expect_null(p$labels$ylab, NULL)
+    expect_snapshot(p$labels)
   }
 )
 
@@ -96,17 +59,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-
-    expect_identical(
-      pb$plot$plot_env$legend.title,
-      ggplot2::expr(
-        atop(atop(scriptstyle(bold("sample sizes:")), italic(n) ~
-        "=" ~ "11"), atop(
-          scriptstyle(bold("correlation (partial):")),
-          "Pearson (Winsorized)"
-        ))
-      )
-    )
+    expect_snapshot(p$labels)
   }
 )
 
@@ -137,16 +90,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-
-    # checking plot labels
-    expect_identical(
-      pb$plot$plot_env$legend.title,
-      ggplot2::expr(atop(atop(scriptstyle(bold("sample sizes:")), italic(n) ~
-      "=" ~ "30"), atop(
-        scriptstyle(bold("correlation (partial):")),
-        "Pearson (Winsorized)"
-      )))
-    )
+    expect_snapshot(p$labels)
   }
 )
 
@@ -180,23 +124,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-
-    expect_identical(
-      pb$plot$plot_env$legend.title,
-      ggplot2::expr(atop(atop(
-        atop(scriptstyle(bold("sample sizes:")), italic(n)[min] ~
-        "=" ~ "32"),
-        atop(
-          italic(n)[mode] ~ "=" ~ "32",
-          italic(n)[max] ~ "=" ~ "83"
-        )
-      ), atop(
-        scriptstyle(
-          bold("correlation:")
-        ),
-        "Spearman"
-      )))
-    )
+    expect_snapshot(p$labels)
   }
 )
 
@@ -211,20 +139,10 @@ test_that(
     p <- suppressWarnings(ggcorrmat(ggplot2::msleep, type = "bayes"))
     pb <- ggplot2::ggplot_build(p)
 
-    # check legend
-    expect_identical(
-      p$plot_env$legend.title,
-      ggplot2::expr(atop(atop(atop(scriptstyle(bold("sample sizes:")), italic(n)[min] ~
[…]
