### 2021-04-21 IndrajeetPatil/ggstatsplot#579 @kevinrue
Brilliant! I think it looks great now. I hope you agree as the developer, but as a user I find it both clearer to spot the X's and more readable (the numbers behind the X's).
