### 2021-04-25 afb63b7
faq; more exaples to ggcoefstats
