+++ R/ggbetweenstats.R
@@ -134,29 +134,25 @@
 #' library(ggstatsplot)
 #'
 #' # simple function call with the defaults
-#' ggstatsplot::ggbetweenstats(
+#' ggbetweenstats(
 #'   data = mtcars,
 #'   x = am,
-#'   y = mpg,
-#'   title = "Fuel efficiency by type of car transmission",
-#'   caption = "Transmission (0 = automatic, 1 = manual)"
+#'   y = mpg
 #' )
 #'
 #' # more detailed function call
-#' ggstatsplot::ggbetweenstats(
-#'   data = datasets::morley,
+#' ggbetweenstats(
+#'   data = morley,
 #'   x = Expt,
 #'   y = Speed,
-#'   type = "nonparametric",
+#'   type = "robust",
 #'   plot.type = "box",
 #'   xlab = "The experiment number",
 #'   ylab = "Speed-of-light measurement",
 #'   pairwise.comparisons = TRUE,
 #'   p.adjust.method = "fdr",
 #'   outlier.tagging = TRUE,
-#'   outlier.label = Run,
-#'   ggtheme = ggplot2::theme_grey(),
-#'   ggstatsplot.layer = FALSE
+#'   outlier.label = Run
 #' )
 #' }
 #' @export
