### 2021-04-26 7647eb6
use more snapshots; refactor check_outliers
