+++ tests/testthat/test-ggpiestats.R
@@ -11,7 +11,6 @@ test_that(
       ggpiestats(
         data = ggplot2::msleep,
         x = vore,
-        bf.message = TRUE,
         title = "mammalian sleep",
         legend.title = "vore",
         caption = "From ggplot2 package",
@@ -125,14 +124,11 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(list(pb$data, pb1$data))
+    expect_snapshot(within(pb$plot$labels, rm(subtitle, caption)))
 
     # checking plot labels
     expect_identical(pb$plot$labels$subtitle, p_subtitle)
     expect_identical(pb$plot$labels$caption, p_cap)
-    expect_null(pb$plot$labels$x, NULL)
-    expect_null(pb$plot$labels$y, NULL)
-    expect_identical(pb$plot$guides$fill$title[1], "transmission")
-    expect_type(pb1$plot$labels$subtitle, "language")
   }
 )
 
@@ -170,26 +166,12 @@ test_that(
     # build the plot
     pb <- ggplot2::ggplot_build(p)
 
-    # checking data used to create a plot
-    dat <- p$data %>%
-      dplyr::mutate_if(
-        .tbl = .,
-        .predicate = is.factor,
-        .funs = ~ as.character(.)
-      )
-
-    # testing everything is okay with data
-    expect_equal(dim(dat), c(4L, 5L))
-    expect_equal(dat$perc, c(8.46, 48.38, 91.54, 51.62), tolerance = 1e-3)
-    expect_equal(dat$Survived[1], "No")
-    expect_equal(dat$Survived[4], "Yes")
-    expect_equal(dat$Sex[2], "Female")
-    expect_equal(dat$Sex[3], "Male")
-    expect_identical(dat$counts, c(126L, 344L, 1364L, 367L))
+    # checking data
+    expect_snapshot(pb$data)
+    expect_snapshot(within(pb$plot$labels, rm(subtitle)))
 
     # checking plot labels
     expect_identical(pb$plot$labels$subtitle, p_subtitle)
-    expect_null(pb$plot$labels$caption, NULL)
   }
 )
 
@@ -237,28 +219,9 @@ test_that(
       )$expression[[1]]
 
     # checking plot labels
-    expect_identical(
-      pb$plot$labels,
-      list(
-        x = NULL,
-        y = NULL,
-        title = NULL,
-        subtitle = p_subtitle,
-        caption = NULL,
-        fill = "1st survey",
-        label = ".label",
-        group = "1st survey"
-      )
-    )
-
-    # labels
-    expect_identical(
-      pb$data[[3]]$label,
-      c(
-        "list(~chi['gof']^2~(1)==569.62, ~italic(p)=='6.8e-126', ~italic(n)==880)",
-        "list(~chi['gof']^2~(1)==245.00, ~italic(p)=='3.2e-55', ~italic(n)==720)"
-      )
-    )
+    expect_snapshot(within(pb$plot$labels, rm(subtitle)))
+    expect_identical(pb$plot$labels$subtitle, p_subtitle)
+    expect_snapshot(pb$data[[3]])
   }
 )
 
@@ -325,20 +288,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-
-    expect_equal(
-      pb$plot$labels,
-      list(
-        x = NULL,
-        y = NULL,
-        title = NULL,
-        subtitle = NULL,
-        caption = NULL,
-        fill = "mode",
-        label = ".label",
-        group = "mode"
-      )
-    )
+    expect_snapshot(pb$plot$labels)
   }
 )
 
