+++ tests/testthat/test-pairwise_ggsignif.R
@@ -8,7 +8,7 @@ test_that(
     # creating the plot
     set.seed(123)
     p <-
-      ggstatsplot::ggbetweenstats(
+      ggbetweenstats(
         data = ggplot2::msleep,
         x = vore,
         y = brainwt,
@@ -25,20 +25,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data[[6]])
-
-    # checking caption
-    expect_identical(
-      pb$plot$labels$caption,
-      ggplot2::expr(atop(
-        displaystyle("mammalian sleep"),
-        expr = paste(
-          "Pairwise test: ",
-          bold("Games-Howell test"),
-          "; Comparisons shown: ",
-          bold("only significant")
-        )
-      ))
-    )
+    expect_snapshot(pb$plot$labels)
   }
 )
 
@@ -53,8 +40,8 @@ test_that(
     # creating the plot
     set.seed(123)
     p <-
-      ggstatsplot::ggbetweenstats(
-        data = ggstatsplot::movies_wide,
+      ggbetweenstats(
+        data = movies_wide,
         x = mpaa,
         y = votes,
         results.subtitle = FALSE,
@@ -71,20 +58,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data[[6]])
-
-    # checking caption
-    expect_identical(
-      pb$plot$labels$caption,
-      ggplot2::expr(atop(
-        displaystyle(NULL),
-        expr = paste(
-          "Pairwise test: ",
-          bold("Games-Howell test"),
-          "; Comparisons shown: ",
-          bold("only non-significant")
-        )
-      ))
-    )
+    expect_snapshot(pb$plot$labels)
   }
 )
 
@@ -98,9 +72,9 @@ test_that(
     # creating the plot
     set.seed(123)
     p <-
-      ggstatsplot::ggbetweenstats(
+      ggbetweenstats(
         data = dplyr::filter(
-          ggstatsplot::movies_long,
+          movies_long,
           genre %in% c("Action", "Comedy", "RomCom")
         ),
         x = genre,
@@ -119,47 +93,8 @@ test_that(
 
     # check data
     set.seed(123)
-    expect_snapshot(pb$data[[6]])
-
-    # data used for pairwise comparisons
-    dat <- pb$plot$plot_env$df_pairwise
-
-    # ggsignif layer parameters
-    ggsignif_stat <- pb$plot$layers[[7]]$stat_params
-
-    # checking dimensions of data
-    data_dims <- dim(dat)
-
-    # data for geom_signif layer
-    data_signif <- as_tibble(pb$data[[7]])
-
-    # checking comparison groups and labels
-    expect_identical(
-      pb$plot$labels$caption$expr,
-      ggplot2::expr(paste(
-        "Pairwise test: ",
-        bold("Dunn test"),
-        "; Comparisons shown: ",
-        bold("all")
-      ))
-    )
-
-    # checking values
-    expect_equal(dat$statistic,
-      c(0.238306686320387, 3.63442246865882, 3.69050171235682),
-      tolerance = 0.001
-    )
-
-    # checking ggsignif layers
-    expect_equal(ggsignif_stat$y_position,
-      c(9.2250, 9.7725, 10.3200),
-      tolerance = 0.001
-    )
-    expect_equal(
-      ggsignif_stat$comparisons[[1]],
-      c("Action", "Comedy")
-    )
-    expect_equal(ggsignif_stat$annotations, dat$label)
+    expect_snapshot(list(pb$data[[6]], pb$data[[7]]))
+    expect_snapshot(pb$plot$labels)
   }
 )
 
[…]

[…]
