+++ R/ggwithinstats.R
@@ -4,8 +4,6 @@
 #'
 #' @description
 #'
-#'
-#'
 #' A combination of box and violin plots along with raw (unjittered) data points
 #' for within-subjects designs with statistical details included in the plot as
 #' a subtitle.
