+++ tests/testthat/test-vdiffr_figs.R
@@ -110,18 +110,20 @@ if (getRversion() < "4.1" && require("vdiffr")) {
 
     ## ----ggcorrmat------------------------------------
 
-    set.seed(123)
-    vdiffr::expect_doppelganger(
-      title = "ggcorrmat works",
-      fig = ggcorrmat(iris)
-    )
+    if (require("ggcorrplot")) {
+      set.seed(123)
+      vdiffr::expect_doppelganger(
+        title = "ggcorrmat works",
+        fig = ggcorrmat(iris)
+      )
 
 
-    set.seed(123)
-    vdiffr::expect_doppelganger(
-      title = "grouped_ggcorrmat works",
-      fig = grouped_ggcorrmat(iris, grouping.var = Species)
-    )
+      set.seed(123)
+      vdiffr::expect_doppelganger(
+        title = "grouped_ggcorrmat works",
+        fig = grouped_ggcorrmat(iris, grouping.var = Species)
+      )
+    }
 
     ## ----ggpiestats-----------------------------------
 

[…]
