### 2021-04-26 2f3adff
improve readability of conditional tests
