+++ tests/testthat/test-ggcorrmat.R
@@ -1,174 +1,178 @@
-if (require("ggcorrplot")) {
 
-  # pearson's r without NAs ------------------------------------------------
+# pearson's r without NAs ------------------------------------------------
 
-  test_that(
-    desc = "checking ggcorrmat - without NAs - pearson's r",
-    code = {
-      skip_on_cran()
+test_that(
+  desc = "checking ggcorrmat - without NAs - pearson's r",
+  code = {
+    skip_on_cran()
+    skip_if_not_installed("ggcorrplot")
 
-      # creating the plot
-      set.seed(123)
-      p <-
-        ggcorrmat(
-          data = iris,
-          cor.vars.names = "x",
-          type = "p",
-          title = "Iris dataset",
-          subtitle = "By Edgar Anderson",
-          ggstatsplot.layer = FALSE,
-          sig.level = 0.001,
-          matrix.type = "full",
-          p.adjust.method = "fdr",
-          colors = NULL,
-          k = 4,
-          ggcorrplot.args = list(
-            lab_col = "white",
-            pch.col = "white"
-          )
+    # creating the plot
+    set.seed(123)
+    p <-
+      ggcorrmat(
+        data = iris,
+        cor.vars.names = "x",
+        type = "p",
+        title = "Iris dataset",
+        subtitle = "By Edgar Anderson",
+        ggstatsplot.layer = FALSE,
+        sig.level = 0.001,
+        matrix.type = "full",
+        p.adjust.method = "fdr",
+        colors = NULL,
+        k = 4,
+        ggcorrplot.args = list(
+          lab_col = "white",
+          pch.col = "white"
         )
+      )
 
-      # checking legend title
-      pb <- ggplot2::ggplot_build(p)
-      p_legend_title <- pb$plot$plot_env$legend.title
+    # checking legend title
+    pb <- ggplot2::ggplot_build(p)
+    p_legend_title <- pb$plot$plot_env$legend.title
 
-      # check data
-      set.seed(123)
-      expect_snapshot(pb$data)
-      expect_snapshot(p$labels)
-    }
-  )
+    # check data
+    set.seed(123)
+    expect_snapshot(pb$data)
+    expect_snapshot(p$labels)
+  }
+)
 
-  # robust r without NAs ---------------------------------------------------
+# robust r without NAs ---------------------------------------------------
 
-  test_that(
-    desc = "checking ggcorrmat - without NAs - robust r",
-    code = {
-      skip_on_cran()
+test_that(
+  desc = "checking ggcorrmat - without NAs - robust r",
+  code = {
+    skip_on_cran()
+    skip_if_not_installed("ggcorrplot")
 
-      # creating the plot
-      set.seed(123)
-      p <-
-        ggcorrmat(
-          data = anscombe,
-          type = "r",
-          partial = TRUE,
-          cor.vars.names = names(anscombe)
-        )
+    # creating the plot
+    set.seed(123)
+    p <-
+      ggcorrmat(
+        data = anscombe,
+        type = "r",
+        partial = TRUE,
+        cor.vars.names = names(anscombe)
+      )
 
-      pb <- ggplot2::ggplot_build(p)
+    pb <- ggplot2::ggplot_build(p)
 
-      # check data
-      set.seed(123)
-      expect_snapshot(pb$data)
-      expect_snapshot(p$labels)
-    }
-  )
+    # check data
+    set.seed(123)
+    expect_snapshot(pb$data)
+    expect_snapshot(p$labels)
+  }
+)
 
-  # robust r with NAs ---------------------------------------------------
[…]
