+++ tests/testthat/test-ggscatterstats.R
@@ -1,351 +1,352 @@
-if (require("ggExtra")) {
-
-  # pearson's r with NAs ---------------------------------------------
-
-  test_that(
-    desc = "checking ggscatterstats - without NAs - pearson's r",
-    code = {
-      skip_on_cran()
-
-      # creating the plot
-      set.seed(123)
-      p <-
-        ggscatterstats(
-          data = dplyr::filter(ggplot2::msleep, conservation == "lc"),
-          x = sleep_total,
-          y = "sleep_cycle",
-          label.var = "name",
-          label.expression = sleep_cycle > 0.3,
-          xlab = "sleep (total)",
-          ylab = "sleep cycle",
-          type = "p",
-          xfill = "red",
-          yfill = "orange",
-          marginal = FALSE,
-          caption = "ggplot2 dataset",
-          title = "Mammalian sleep"
-        )
-
-      # plot build
-      pb <- ggplot2::ggplot_build(p)
-
-      # check data
-      set.seed(123)
-      expect_snapshot(list(pb$data[[1]], head(pb$data[[2]]), pb$data[[3]]))
-
-      # subtitle
-      set.seed(123)
-      p_subtitle <-
-        statsExpressions::corr_test(
-          data = dplyr::filter(ggplot2::msleep, conservation == "lc"),
-          x = "sleep_total",
-          y = sleep_cycle,
-          type = "p"
-        )$expression[[1]]
-
-      # subtitle
-      set.seed(123)
-      p_cap <-
-        statsExpressions::corr_test(
-          data = dplyr::filter(ggplot2::msleep, conservation == "lc"),
-          x = "sleep_total",
-          y = sleep_cycle,
-          top.text = "ggplot2 dataset",
-          type = "bayes"
-        )$expression[[1]]
-
-      # checking plot labels
-      expect_identical(pb$plot$labels$caption, p_cap)
-      expect_identical(pb$plot$labels$subtitle, p_subtitle)
-      expect_snapshot(within(pb$plot$labels, rm(subtitle, caption)))
-    }
-  )
-
-  # spearman's rho with NAs ---------------------------------------------
-
-  test_that(
-    desc = "checking ggscatterstats - without NAs - spearman's rho",
-    code = {
-      skip_on_cran()
-
-      # creating the plot
-      set.seed(123)
-      p <-
-        ggscatterstats(
-          data = dplyr::filter(ggplot2::msleep, conservation == "lc"),
-          x = "sleep_total",
-          y = sleep_cycle,
-          type = "np",
-          conf.level = 0.99,
-          marginal = FALSE
-        )
-
-      # build the plot
-      pb <- ggplot2::ggplot_build(p)
-
-      # check data
-      set.seed(123)
-      expect_snapshot(pb$data[[1]])
-
-      # subtitle
-      set.seed(123)
-      p_subtitle <-
-        statsExpressions::corr_test(
-          data = dplyr::filter(ggplot2::msleep, conservation == "lc"),
-          x = sleep_total,
-          y = sleep_cycle,
-          type = "np",
-          conf.level = 0.99
-        )$expression[[1]]
-
-      # testing data and annotations
-      expect_identical(pb$plot$labels$subtitle, p_subtitle)
-      expect_snapshot(within(pb$plot$labels, rm(subtitle)))
-    }
-  )
 
 
-  # winsorized Pearson with NAs ---------------------------------------------
+# pearson's r with NAs ---------------------------------------------
 
-  test_that(
-    desc = "checking ggscatterstats - without NAs - winsorized Pearson",
-    code = {
-      skip_on_cran()
+test_that(
[…]

[…]
