+++ R/grouped_ggdotplotstats.R
@@ -3,8 +3,6 @@
 #'
 #' @description
 #'
-#'
-#'
 #' Helper function for `ggstatsplot::ggdotplotstats` to apply this function
 #' across multiple levels of a given factor and combining the resulting plots
 #' using `ggstatsplot::combine_plots`.
@@ -30,7 +28,7 @@
 #' library(ggstatsplot)
 #'
 #' # removing factor level with very few no. of observations
-#' df <- dplyr::filter(.data = ggplot2::mpg, cyl %in% c("4", "6", "8"))
+#' df <- dplyr::filter(ggplot2::mpg, cyl %in% c("4", "6", "8"))
 #'
 #' # plot
 #' grouped_ggdotplotstats(
