### 2021-04-29 9a98c04
clean up other resources vignette
