### 2021-04-29 ad64c5b
add shiny faq
