+++ paper/paper.Rmd
@@ -11,11 +11,9 @@ tags:
 authors:
   - name: Indrajeet Patil
     orcid: 0000-0003-1995-6531
-    affiliation: 1
 affiliations:
   - name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
-citation_author: Patil
-date: "`r Sys.Date()`"
+    index: 1
 bibliography: paper.bib
 output: rticles::joss_article
 csl: apa.csl
@@ -62,7 +60,7 @@ phases into one in the form of an informative graphic with statistical details.
 Before discussing benefits of this approach, we will see one example (Figure
 1).
 
-```{r penguins, fig.width=8, fig.height=5, fig.cap="Example plot from the `ggstatsplot` package illustrates its philosophy of juxtaposing informative visualizations with details from statistical analysis. To see all supported plots and statistical analyses, see the package website: https://indrajeetpatil.github.io/ggstatsplot/"}
+```{r penguins, fig.width=8, fig.height=5, fig.cap="Example plot from the `ggstatsplot` package illustrates its philosophy of juxtaposing informative visualizations with details from statistical analysis. To see all supported plots and statistical analyses, see the package website: \\url{https://indrajeetpatil.github.io/ggstatsplot/}"}
 set.seed(123) # for reproducibility
 library(palmerpenguins) # for 'penguins' dataset
 library(ggstatsplot)
