### 2021-05-01 e945b34
simplify grouped functions
