+++ R/grouped_gghistostats.R
@@ -72,11 +72,11 @@ grouped_gghistostats <- function(data,
 
   # getting the dataframe ready
   df <-
-    dplyr::select(.data = data, {{ grouping.var }}, {{ x }}) %>%
+    dplyr::select(data, {{ grouping.var }}, {{ x }}) %>%
     grouped_list(grouping.var = {{ grouping.var }})
 
   # creating a list of plots
-  plotlist_purrr <-
+  p_ls <-
     purrr::pmap(
       .l = list(data = df, title = names(df)),
       .f = ggstatsplot::gghistostats,
@@ -88,9 +88,8 @@ grouped_gghistostats <- function(data,
     )
 
   # combining the list of plots into a single plot
-  if (output == "plot") {
-    return(combine_plots(plotlist_purrr, plotgrid.args = plotgrid.args, annotation.args = annotation.args))
-  } else {
-    return(plotlist_purrr) # subtitle list
-  }
+  if (output == "plot") p_ls <- combine_plots(p_ls, plotgrid.args, annotation.args)
+
+  # return the object
+  p_ls
 }
