+++ R/grouped_ggpiestats.R
@@ -46,13 +46,13 @@ grouped_ggpiestats <- function(data,
 
   # creating a dataframe
   df <-
-    dplyr::select(.data = data, {{ grouping.var }}, {{ x }}, {{ y }}, {{ counts }}) %>%
+    dplyr::select(data, {{ grouping.var }}, {{ x }}, {{ y }}, {{ counts }}) %>%
     grouped_list(grouping.var = {{ grouping.var }})
 
   # ==================== creating a list of return objects ===================
 
   # creating a list of plots using `pmap`
-  plotlist_purrr <-
+  p_ls <-
     purrr::pmap(
       .l = list(data = df, title = names(df)),
       .f = ggstatsplot::ggpiestats,
@@ -65,9 +65,8 @@ grouped_ggpiestats <- function(data,
     )
 
   # combining the list of plots into a single plot
-  if (output == "plot") {
-    return(combine_plots(plotlist_purrr, plotgrid.args = plotgrid.args, annotation.args = annotation.args))
-  } else {
-    return(plotlist_purrr) # subtitle list
-  }
+  if (output == "plot") p_ls <- combine_plots(p_ls, plotgrid.args, annotation.args)
+
+  # return the object
+  p_ls
 }
