+++ R/ggbarstats.R
@@ -20,7 +20,6 @@
 #'
 #' @importFrom dplyr select mutate
 #' @importFrom rlang !!! as_name ensym exec
-#' @importFrom paletteer scale_fill_paletteer_d
 #' @importFrom tidyr uncount drop_na
 #' @importFrom statsExpressions contingency_table
 #'
