### 2021-05-05 #3167 @csoneson
@IndrajeetPatil - let's have some patience, these are challenging times for everyone (in particular, note that JOSS is still operating in [reduced service mode](https://blog.joss.theoj.org/2020/05/reopening-joss) because of the challenges posed by the pandemic).
