### 2021-05-05 #3167 @IndrajeetPatil
Oh, I see. Thanks for pointing out the reduced service mode, which does state:

> "In this reduced service mode, we’re asking that reviewers try to complete their reviews in six weeks."

So happy to wait for 3 more weeks and revisit this.
