### 2021-05-08 7427c05
add point.args to ggwithinstats; minor doc changes
