+++ R/grouped_ggdotplotstats.R
@@ -54,20 +54,18 @@ grouped_ggdotplotstats <- function(data,
   # ======================== preparing dataframe ============================
 
   # creating a dataframe
-  df <-
-    dplyr::select(data, {{ grouping.var }}, {{ x }}, {{ y }}) %>%
+  df <- dplyr::select(data, {{ grouping.var }}, {{ x }}, {{ y }}) %>%
     grouped_list(grouping.var = {{ grouping.var }})
 
   # creating a list of plots
-  p_ls <-
-    purrr::pmap(
-      .l = list(data = df, title = names(df)),
-      .f = ggstatsplot::ggdotplotstats,
-      x = {{ x }},
-      y = {{ y }},
-      output = output,
-      ...
-    )
+  p_ls <- purrr::pmap(
+    .l = list(data = df, title = names(df)),
+    .f = ggstatsplot::ggdotplotstats,
+    x = {{ x }},
+    y = {{ y }},
+    output = output,
+    ...
+  )
 
   # combining the list of plots into a single plot
   if (output == "plot") p_ls <- combine_plots(p_ls, plotgrid.args, annotation.args)
