+++ R/ggcoefstats.R
@@ -188,15 +188,14 @@ ggcoefstats <- function(x,
     if (effsize == "omega") omega_squared <- "partial"
 
     # converting model object to a tidy dataframe
-    tidy_df <-
-      parameters::model_parameters(
-        model = x,
-        eta_squared = eta_squared,
-        omega_squared = omega_squared,
-        ci = conf.level,
-        verbose = FALSE,
-        ...
-      ) %>%
+    tidy_df <- parameters::model_parameters(
+      model = x,
+      eta_squared = eta_squared,
+      omega_squared = omega_squared,
+      ci = conf.level,
+      verbose = FALSE,
+      ...
+    ) %>%
       parameters::standardize_names(style = "broom") %>%
       dplyr::rename_all(~ gsub("omega2.|eta2.", "", .x))
 
@@ -317,15 +316,14 @@ ggcoefstats <- function(x,
     # if glance is not available, inform the user
     if (!is.null(glance_df) && all(c("aic", "bic") %in% names(glance_df))) {
       # preparing caption with model diagnostics
-      caption <-
-        substitute(
-          expr = atop(displaystyle(top.text), expr = paste("AIC = ", AIC, ", BIC = ", BIC)),
-          env = list(
-            top.text = caption,
-            AIC = format_value(glance_df$aic[[1]], 0L),
-            BIC = format_value(glance_df$bic[[1]], 0L)
-          )
+      caption <- substitute(
+        expr = atop(displaystyle(top.text), expr = paste("AIC = ", AIC, ", BIC = ", BIC)),
+        env = list(
+          top.text = caption,
+          AIC = format_value(glance_df$aic[[1]], 0L),
+          BIC = format_value(glance_df$bic[[1]], 0L)
         )
+      )
     }
   }
 
@@ -341,13 +339,12 @@ ggcoefstats <- function(x,
 
     # results from Bayesian random-effects meta-analysis (only for parametric)
     if (meta.type == "parametric" && isTRUE(bf.message)) {
-      caption_df <-
-        statsExpressions::meta_analysis(
-          top.text = caption,
-          type = "bayes",
-          data = tidy_df,
-          k = k
-        )
+      caption_df <- statsExpressions::meta_analysis(
+        top.text = caption,
+        type = "bayes",
+        data = tidy_df,
+        k = k
+      )
 
       caption <- caption_df$expression[[1]]
     }
@@ -359,13 +356,12 @@ ggcoefstats <- function(x,
   tidy_df %<>% dplyr::mutate(term = as.factor(term), .rowid = dplyr::row_number())
 
   # sorting factor levels
-  new_order <-
-    switch(sort,
-      "none" = order(tidy_df$.rowid, decreasing = FALSE),
-      "ascending" = order(tidy_df$estimate, decreasing = FALSE),
-      "descending" = order(tidy_df$estimate, decreasing = TRUE),
-      order(tidy_df$.rowid, decreasing = FALSE)
-    )
+  new_order <- switch(sort,
+    "none" = order(tidy_df$.rowid, decreasing = FALSE),
+    "ascending" = order(tidy_df$estimate, decreasing = FALSE),
+    "descending" = order(tidy_df$estimate, decreasing = TRUE),
+    order(tidy_df$.rowid, decreasing = FALSE)
+  )
 
   # sorting `term` factor levels according to new sorting order
   tidy_df %<>%
@@ -378,7 +374,7 @@ ggcoefstats <- function(x,
   # palette check is necessary only if output is a plot
   if (output == "plot") {
     # setting up the basic architecture
-    plot <- ggplot2::ggplot(data = tidy_df, mapping = ggplot2::aes(x = estimate, y = term))
+    plot <- ggplot2::ggplot(tidy_df, mapping = ggplot2::aes(x = estimate, y = term))
 
[…]
