### 2021-05-11 e2dd95d
fix visual regression tests for statsExpr update
