+++ DESCRIPTION
@@ -39,7 +39,7 @@ Imports:
     purrr,
     rlang,
     stats,
-    statsExpressions (>= 1.0.1),
+    statsExpressions (>= 1.0.1.9000),
     tidyr,
     utils
 Suggests:
@@ -57,6 +57,8 @@ Suggests:
     tibble,
     vdiffr,
     WRS2
+Remotes:
+    IndrajeetPatil/statsExpressions
 VignetteBuilder: 
     knitr
 Encoding: UTF-8
