+++ tests/testthat/test-ggwithinstats.R
@@ -142,40 +142,6 @@ if (require("afex")) {
         expect_snapshot(within(pb1$plot$labels, rm(subtitle, caption)))
 
         expect_identical(p1$labels$subtitle, p1_subtitle)
-        # expect_identical(
-        #   pb1$plot$labels$caption,
-        #   ggplot2::expr(atop(
-        #     displaystyle(atop(
-        #       displaystyle("From `WRS2` package"),
-        #       expr = paste(
-        #         "log"["e"],
-        #         "(BF"["01"],
-        #         ") = ",
-        #         "-2.1154",
-        #         ", ",
-        #         widehat(italic(R^"2"))["median"]^"posterior",
-        #         " = ",
-        #         "0.8930",
-        #         ", CI"["95%"]^"HDI",
-        #         " [",
-        #         "0.8499",
-        #         ", ",
-        #         "0.9228",
-        #         "]",
-        #         ", ",
-        #         italic("r")["Cauchy"]^"JZS",
-        #         " = ",
-        #         "0.7070"
-        #       )
-        #     )),
-        #     expr = paste(
-        #       "Pairwise test: ",
-        #       bold("Student's t-test"),
-        #       "; Comparisons shown: ",
-        #       bold("only significant")
-        #     )
-        #   ))
-        # )
       }
     }
   )
