+++ tests/testthat/test-vdiffr_figs.R
@@ -1,4 +1,4 @@
-if (getRversion() < "4.1") {
+if (getRversion() < "4.1" && packageVersion("statsExpressions") >= "1.0.1.9000") {
   test_that("plots are rendered correctly", {
     skip_on_cran()
     skip_if_not_installed("vdiffr")
