+++ R/reexports.R
@@ -27,3 +27,8 @@ ipmisc::enframe
 #' @export
 #' @importFrom ipmisc as_tibble
 ipmisc::as_tibble
+
+
+#' @export
+#' @importFrom effectsize effectsize
+effectsize::effectsize
