+++ man/reexports.Rd
@@ -10,6 +10,7 @@
 \alias{tibble}
 \alias{enframe}
 \alias{as_tibble}
+\alias{effectsize}
 \title{Objects exported from other packages}
 \keyword{internal}
 \description{
@@ -17,6 +18,8 @@ These objects are imported from other packages. Follow the links
 below to see their documentation.
 
 \describe{
+  \item{effectsize}{\code{\link[effectsize]{effectsize}}}
+
   \item{ipmisc}{\code{\link[ipmisc:reexports]{\%$\%}}, \code{\link[ipmisc:reexports]{\%<-\%}}, \code{\link[ipmisc:reexports]{\%<>\%}}, \code{\link[ipmisc:reexports]{\%>\%}}, \code{\link[ipmisc:reexports]{as_tibble}}, \code{\link[ipmisc:reexports]{enframe}}, \code{\link[ipmisc:reexports]{tibble}}}
 }}
 
