+++ NAMESPACE
@@ -6,6 +6,7 @@ export("%<>%")
 export("%>%")
 export(as_tibble)
 export(combine_plots)
+export(effectsize)
 export(enframe)
 export(ggbarstats)
 export(ggbetweenstats)
@@ -52,6 +53,7 @@ importFrom(dplyr,summarize)
 importFrom(dplyr,tally)
 importFrom(dplyr,ungroup)
 importFrom(dplyr,vars)
+importFrom(effectsize,effectsize)
 importFrom(ggrepel,geom_label_repel)
 importFrom(ggsignif,geom_signif)
 importFrom(insight,find_statistic)
