+++ tests/testthat/test-z_ggcoefstats_dataframes.R
@@ -1,5 +1,7 @@
 # dataframe as input ----------------------------------------------------
 
+# should be run last
+
 if (all(unlist(lapply(
   c("metafor", "metaBMA"), # needed packages
   require,
