+++ DESCRIPTION
@@ -70,3 +70,4 @@ Roxygen: list(markdown = TRUE, roclets = c("rd", "namespace", "collate",
     "pkgapi::api_roclet"))
 RoxygenNote: 7.1.1.9001
 Config/testthat/edition: 3
+Config/testthat/parallel: true

[…]
