### 2021-05-17 e3dacf2
improve mcp snapshot tests
