+++ R/combine_plots.R
@@ -28,20 +28,18 @@
 #' library(ggplot2)
 #'
 #' # preparing the first plot
-#' p1 <-
-#'   ggplot2::ggplot(
-#'     data = subset(iris, iris$Species == "setosa"),
-#'     aes(x = Sepal.Length, y = Sepal.Width)
-#'   ) +
+#' p1 <- ggplot2::ggplot(
+#'   data = subset(iris, iris$Species == "setosa"),
+#'   aes(x = Sepal.Length, y = Sepal.Width)
+#' ) +
 #'   geom_point() +
 #'   labs(title = "setosa")
 #'
 #' # preparing the second plot
-#' p2 <-
-#'   ggplot2::ggplot(
-#'     data = subset(iris, iris$Species == "versicolor"),
-#'     aes(x = Sepal.Length, y = Sepal.Width)
-#'   ) +
+#' p2 <- ggplot2::ggplot(
+#'   data = subset(iris, iris$Species == "versicolor"),
+#'   aes(x = Sepal.Length, y = Sepal.Width)
+#' ) +
 #'   geom_point() +
 #'   labs(title = "versicolor")
 #'
