+++ R/ggwithinstats.R
@@ -271,7 +271,7 @@ ggwithinstats <- function(data,
 
   if (isTRUE(pairwise.comparisons) && test == "anova") {
     # creating dataframe with pairwise comparison results
-    df_pairwise <- pairwiseComparisons::pairwise_comparisons(
+    df_mcp <- pairwiseComparisons::pairwise_comparisons(
       data = data,
       x = {{ x }},
       y = {{ y }},
@@ -285,7 +285,7 @@ ggwithinstats <- function(data,
     # adding the layer for pairwise comparisons
     plot <- ggsignif_adder(
       plot = plot,
-      df_pairwise = df_pairwise,
+      df_mcp = df_mcp,
       data = data,
       x = {{ x }},
       y = {{ y }},
@@ -297,7 +297,7 @@ ggwithinstats <- function(data,
     if (type != "bayes") {
       caption <- pairwiseComparisons::pairwise_caption(
         caption,
-        unique(df_pairwise$test.details),
+        unique(df_mcp$test.details),
         pairwise.display
       )
     }
