+++ R/helpers_ggbetweenstats_graphics.R
@@ -39,9 +39,6 @@ centrality_ggrepel <- function(plot,
                                data,
                                x,
                                y,
-                               type = "parametric",
-                               tr = 0.2,
-                               k = 2L,
                                centrality.path = FALSE,
                                centrality.path.args = list(
                                  color = "red",
@@ -56,11 +53,7 @@ centrality_ggrepel <- function(plot,
                                ),
                                ...) {
   # creating the dataframe
-  centrality_df <- centrality_data(data, {{ x }}, {{ y }},
-    type = type,
-    tr = tr,
-    k = k
-  )
+  centrality_df <- centrality_data(data, {{ x }}, {{ y }}, ...)
 
   # if there should be lines connecting mean values across groups
   if (isTRUE(centrality.path)) {
@@ -73,8 +66,6 @@ centrality_ggrepel <- function(plot,
         !!!centrality.path.args
       )
   }
-
-  # ------------------------ plot -------------------------------------
 
   # highlight the mean of each group
   plot +
@@ -150,7 +141,7 @@ centrality_data <- function(data,
 #'
 #' @param ... Currently ignored.
 #' @param plot A `ggplot` object on which `geom_signif` needed to be added.
-#' @param df_pairwise A dataframe containing results from pairwise comparisons
+#' @param df_mcp A dataframe containing results from pairwise comparisons
 #'   (produced by `pairwiseComparisons::pairwise_comparisons()` function).
 #' @inheritParams ggbetweenstats
 #'
@@ -180,12 +171,12 @@ centrality_data <- function(data,
 #'   data = iris,
 #'   x = Species,
 #'   y = Sepal.Length,
-#'   df_pairwise = df_pair
+#'   df_mcp = df_pair
 #' )
 #' @noRd
 
 ggsignif_adder <- function(plot,
-                           df_pairwise,
+                           df_mcp,
                            data,
                            x,
                            y,
@@ -193,38 +184,34 @@ ggsignif_adder <- function(plot,
                            ggsignif.args = list(textsize = 3, tip_length = 0.01),
                            ...) {
   # creating a column for group combinations
-  df_pairwise %<>% dplyr::mutate(groups = purrr::pmap(.l = list(group1, group2), .f = c))
+  df_mcp %<>% dplyr::mutate(groups = purrr::pmap(.l = list(group1, group2), .f = c))
 
   # for Bayes Factor, there will be no "p.value" column
-  if ("p.value" %in% names(df_pairwise)) {
-    # decide what needs to be displayed:
-    # only significant comparisons shown
-    if (pairwise.display %in% c("s", "significant")) {
-      df_pairwise %<>% dplyr::filter(p.value < 0.05)
-    }
-
-    # only non-significant comparisons shown
-    if (pairwise.display %in% c("ns", "nonsignificant", "non-significant")) {
-      df_pairwise %<>% dplyr::filter(p.value >= 0.05)
-    }
+  if ("p.value" %in% names(df_mcp)) {
+    # decide what needs to be displayed
+    if (grepl("^s", pairwise.display)) df_mcp %<>% dplyr::filter(p.value < 0.05)
+    if (grepl("^n", pairwise.display)) df_mcp %<>% dplyr::filter(p.value >= 0.05)
 
     # proceed only if there are any significant comparisons to display
-    if (dim(df_pairwise)[[1]] == 0L) {
+    if (dim(df_mcp)[[1]] == 0L) {
       return(plot)
     }
   }
 
   # arrange the dataframe so that annotations are properly aligned
-  df_pairwise %<>% dplyr::arrange(group1, group2)
+  df_mcp %<>% dplyr::arrange(group1, group2)
 
[…]
