+++ tests/testthat/_snaps/ggcorrmat.md
@@ -539,20 +539,20 @@
       [[1]]
             fill x y PANEL group xmin xmax ymin ymax colour size linetype alpha width
       1  #65B997 1 1     1     1  0.5  1.5  0.5  1.5  black  0.1        1    NA    NA
-      2  #FDD49E 1 2     1     2  0.5  1.5  1.5  2.5  black  0.1        1    NA    NA
-      3  #FFE0B9 2 2     1     6  1.5  2.5  1.5  2.5  black  0.1        1    NA    NA
+      2  #FED5A1 1 2     1     2  0.5  1.5  1.5  2.5  black  0.1        1    NA    NA
+      3  #FFE1BB 2 2     1     6  1.5  2.5  1.5  2.5  black  0.1        1    NA    NA
       4  #E69F00 1 3     1     3  0.5  1.5  2.5  3.5  black  0.1        1    NA    NA
       5  #F4B85B 2 3     1     7  1.5  2.5  2.5  3.5  black  0.1        1    NA    NA
-      6  #A4D5BF 3 3     1    10  2.5  3.5  2.5  3.5  black  0.1        1    NA    NA
-      7  #FFDDB2 1 4     1     4  0.5  1.5  3.5  4.5  black  0.1        1    NA    NA
+      6  #A7D6C1 3 3     1    10  2.5  3.5  2.5  3.5  black  0.1        1    NA    NA
+      7  #FFDEB5 1 4     1     4  0.5  1.5  3.5  4.5  black  0.1        1    NA    NA
       8  #FFEBD2 2 4     1     8  1.5  2.5  3.5  4.5  black  0.1        1    NA    NA
-      9  #4FB08B 3 4     1    11  2.5  3.5  3.5  4.5  black  0.1        1    NA    NA
+      9  #52B18C 3 4     1    11  2.5  3.5  3.5  4.5  black  0.1        1    NA    NA
       10 #B9DECD 4 4     1    13  3.5  4.5  3.5  4.5  black  0.1        1    NA    NA
       11 #FFE1BB 1 5     1     5  0.5  1.5  4.5  5.5  black  0.1        1    NA    NA
-      12 #FFDFB7 2 5     1     9  1.5  2.5  4.5  5.5  black  0.1        1    NA    NA
-      13 #AFDAC6 3 5     1    12  2.5  3.5  4.5  5.5  black  0.1        1    NA    NA
+      12 #FFE0B9 2 5     1     9  1.5  2.5  4.5  5.5  black  0.1        1    NA    NA
+      13 #B1DBC8 3 5     1    12  2.5  3.5  4.5  5.5  black  0.1        1    NA    NA
       14 #C1E2D3 4 5     1    14  3.5  4.5  4.5  5.5  black  0.1        1    NA    NA
-      15 #2DA57C 5 5     1    15  4.5  5.5  4.5  5.5  black  0.1        1    NA    NA
+      15 #31A67D 5 5     1    15  4.5  5.5  4.5  5.5  black  0.1        1    NA    NA
          height
       1      NA
       2      NA
@@ -573,36 +573,36 @@
       [[2]]
             fill x y PANEL group colour size angle hjust vjust alpha family fontface
       1  #65B997 1 1     1     1  black    4     0   0.5   0.5    NA               1
-      2  #FDD49E 1 2     1     2  black    4     0   0.5   0.5    NA               1
-      3  #FFE0B9 2 2     1     6  black    4     0   0.5   0.5    NA               1
+      2  #FED5A1 1 2     1     2  black    4     0   0.5   0.5    NA               1
+      3  #FFE1BB 2 2     1     6  black    4     0   0.5   0.5    NA               1
[…]
