+++ tests/testthat/_snaps/ggcorrmat.md
@@ -537,72 +537,16 @@
       pb$data
     Output
       [[1]]
-            fill x y PANEL group xmin xmax ymin ymax colour size linetype alpha width
-      1  #65B997 1 1     1     1  0.5  1.5  0.5  1.5  black  0.1        1    NA    NA
-      2  #FED5A1 1 2     1     2  0.5  1.5  1.5  2.5  black  0.1        1    NA    NA
-      3  #FFE1BB 2 2     1     6  1.5  2.5  1.5  2.5  black  0.1        1    NA    NA
-      4  #E69F00 1 3     1     3  0.5  1.5  2.5  3.5  black  0.1        1    NA    NA
-      5  #F4B85B 2 3     1     7  1.5  2.5  2.5  3.5  black  0.1        1    NA    NA
-      6  #A7D6C1 3 3     1    10  2.5  3.5  2.5  3.5  black  0.1        1    NA    NA
-      7  #FFDEB5 1 4     1     4  0.5  1.5  3.5  4.5  black  0.1        1    NA    NA
-      8  #FFEBD2 2 4     1     8  1.5  2.5  3.5  4.5  black  0.1        1    NA    NA
-      9  #52B18C 3 4     1    11  2.5  3.5  3.5  4.5  black  0.1        1    NA    NA
-      10 #B9DECD 4 4     1    13  3.5  4.5  3.5  4.5  black  0.1        1    NA    NA
-      11 #FFE1BB 1 5     1     5  0.5  1.5  4.5  5.5  black  0.1        1    NA    NA
-      12 #FFE0B9 2 5     1     9  1.5  2.5  4.5  5.5  black  0.1        1    NA    NA
-      13 #B1DBC8 3 5     1    12  2.5  3.5  4.5  5.5  black  0.1        1    NA    NA
-      14 #C1E2D3 4 5     1    14  3.5  4.5  4.5  5.5  black  0.1        1    NA    NA
-      15 #31A67D 5 5     1    15  4.5  5.5  4.5  5.5  black  0.1        1    NA    NA
-         height
-      1      NA
-      2      NA
-      3      NA
-      4      NA
-      5      NA
-      6      NA
-      7      NA
-      8      NA
-      9      NA
-      10     NA
-      11     NA
-      12     NA
-      13     NA
-      14     NA
-      15     NA
+           fill x y PANEL group xmin xmax ymin ymax colour size linetype alpha width
+      1 #2DA57C 1 1     1     1  0.5  1.5  0.5  1.5  black  0.1        1    NA    NA
+        height
+      1     NA
       
       [[2]]
-            fill x y PANEL group colour size angle hjust vjust alpha family fontface
-      1  #65B997 1 1     1     1  black    4     0   0.5   0.5    NA               1
-      2  #FED5A1 1 2     1     2  black    4     0   0.5   0.5    NA               1
-      3  #FFE1BB 2 2     1     6  black    4     0   0.5   0.5    NA               1
-      4  #E69F00 1 3     1     3  black    4     0   0.5   0.5    NA               1
-      5  #F4B85B 2 3     1     7  black    4     0   0.5   0.5    NA               1
-      6  #A7D6C1 3 3     1    10  black    4     0   0.5   0.5    NA               1
-      7  #FFDEB5 1 4     1     4  black    4     0   0.5   0.5    NA               1
-      8  #FFEBD2 2 4     1     8  black    4     0   0.5   0.5    NA               1
-      9  #52B18C 3 4     1    11  black    4     0   0.5   0.5    NA               1
-      10 #B9DECD 4 4     1    13  black    4     0   0.5   0.5    NA               1
[…]

[…]
