### 2021-05-17 9e8673b
add summary of graphics
