+++ R/ggdotplotstats.R
@@ -4,7 +4,7 @@
 #' @description
 #'
 #' A dot chart (as described by William S. Cleveland) with statistical details
-#' from one-sample test included in the plot as a subtitle.
+#' from one-sample test details.
 #'
 #' @param ... Currently ignored.
 #' @param y Label or grouping variable.
@@ -16,7 +16,7 @@
 #' @importFrom dplyr row_number percent_rank pull
 #' @importFrom statsExpressions one_sample_test
 #'
-#' @references
+#' @details For more details, see:
 #' \url{https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/ggdotplotstats.html}
 #'
 #' @seealso \code{\link{grouped_gghistostats}}, \code{\link{gghistostats}},
@@ -57,12 +57,20 @@ ggdotplotstats <- function(data,
                            effsize.type = "g",
                            conf.level = 0.95,
                            tr = 0.2,
-                           k = 2,
+                           k = 2L,
                            results.subtitle = TRUE,
-                           point.args = list(color = "black", size = 3, shape = 16),
+                           point.args = list(
+                             color = "black",
+                             size = 3,
+                             shape = 16
+                           ),
                            centrality.plotting = TRUE,
                            centrality.type = type,
-                           centrality.line.args = list(color = "blue", size = 1),
+                           centrality.line.args = list(
+                             color = "blue",
+                             size = 1,
+                             linetype = "dashed"
+                           ),
                            ggplot.component = NULL,
                            ggtheme = ggplot2::theme_bw(),
                            ggstatsplot.layer = TRUE,
@@ -143,7 +151,7 @@ ggdotplotstats <- function(data,
   # ------------------------------ basic plot ----------------------------
 
   # creating the basic plot
-  plot <- ggplot2::ggplot(data, mapping = ggplot2::aes(x = {{ x }}, y = rank)) +
+  plot <- ggplot2::ggplot(data, mapping = ggplot2::aes({{ x }}, y = rank)) +
     rlang::exec(ggplot2::geom_point, !!!point.args) +
     ggplot2::scale_y_continuous(
       name = ylab,
