+++ README.md
@@ -273,7 +273,7 @@ Note here that the function can be used to tag outliers!
 
 | graphical element        | `geom_` used                | argument for further modification |
 |--------------------------|-----------------------------|-----------------------------------|
-| raw points               | `ggplot2::geom_point`       | `point.args`                      |
+| raw data                 | `ggplot2::geom_point`       | `point.args`                      |
 | box plot                 | `ggplot2::geom_boxplot`     | ❌                                 |
 | density plot             | `ggplot2::geom_violin`      | `violin.args`                     |
 | centrality measure point | `ggplot2::geom_point`       | `centrality.point.args`           |
@@ -410,7 +410,7 @@ grouped_ggwithinstats(
 
 | graphical element             | `geom_` used                | argument for further modification |
 |-------------------------------|-----------------------------|-----------------------------------|
-| raw points                    | `ggplot2::geom_point`       | `point.args`                      |
+| raw data                      | `ggplot2::geom_point`       | `point.args`                      |
 | point path                    | `ggplot2::geom_path`        | `point.path.args`                 |
 | box plot                      | `ggplot2::geom_boxplot`     | ❌                                 |
 | density plot                  | `ggplot2::geom_violin`      | `violin.args`                     |
@@ -528,6 +528,14 @@ grouped_gghistostats(
 
 <img src="man/figures/README-gghistostats2-1.png" width="100%" />
 
+### Summary of graphics
+
+| graphical element       | `geom_` used             | argument for further modification |
+|-------------------------|--------------------------|-----------------------------------|
+| histogram bin           | `ggplot2::stat_bin`      | `bin.args`                        |
+| centrality measure line | `ggplot2::geom_vline`    | `centrality.line.args`            |
+| normality curve         | `ggplot2::stat_function` | `normal.curve.args`               |
+
 ### Summary of tests
 
 **Central tendency measure**
@@ -620,6 +628,13 @@ grouped_ggdotplotstats(
 ```
 
 <img src="man/figures/README-ggdotplotstats2-1.png" width="100%" />
+
+### Summary of graphics
+
+| graphical element       | `geom_` used          | argument for further modification |
+|-------------------------|-----------------------|-----------------------------------|
+| raw data                | `ggplot2::geom_point` | `point.args`                      |
+| centrality measure line | `ggplot2::geom_vline` | `centrality.line.args`            |
 
 ### Summary of tests
 
@@ -741,6 +756,15 @@ grouped_ggscatterstats(
 ```
 
 <img src="man/figures/README-ggscatterstats3-1.png" width="100%" />
+
+### Summary of graphics
+
+| graphical element      | `geom_` used                | argument for further modification |
+|------------------------|-----------------------------|-----------------------------------|
+| raw data               | `ggplot2::geom_point`       | `point.args`                      |
+| labels for raw data    | `ggrepel::geom_label_repel` | `point.label.args`                |
+| smooth line            | `ggplot2::geom_smooth`      | `smooth.line.args`                |
+| marginal distributions | `ggExtra::ggMarginal`       | ❌                                 |
 
 ### Summary of tests
 
@@ -941,6 +965,12 @@ ggcorrmat(
 #> 15 Bayesian Pearson correlation    30
 ```
 
+### Summary of graphics
+
+| graphical element  | `geom_` used             | argument for further modification |
+|--------------------|--------------------------|-----------------------------------|
+| correlation matrix | `ggcorrplot::ggcorrplot` | `ggcorrplot.args`                 |
+
 ### Summary of tests
 
 **Hypothesis testing** and **Effect size estimation**
@@ -1013,6 +1043,13 @@ grouped_ggpiestats(
[…]
