### 2021-05-21 IndrajeetPatil/ggstatsplot#596 @njtierney
JOSS Review: Installation
Running:

```r
remotes::install_github(
  repo = "IndrajeetPatil/ggstatsplot", # package path on GitHub
  dependencies = FALSE, # assumes you have already installed needed packages
  quick = TRUE # skips docs, demos, and vignettes
)
```

resulted in the following error:

```
Using github PAT from envvar GITHUB_PAT
Downloading GitHub repo IndrajeetPatil/ggstatsplot@HEAD
✓  checking for file ‘/private/var/folders/yv/69r1bw2931jb9tkdckxsmkr40000gn/T/Rtmp7q0MRR/remotes15162626b2277/IndrajeetPatil-ggstatsplot-7f4a37a/DESCRIPTION’ ...
─  preparing ‘ggstatsplot’: (463ms)
✓  checking DESCRIPTION meta-information ...
   Removed empty directory ‘ggstatsplot/hextools’
─  building ‘ggstatsplot_0.8.0.tar.gz’

* installing *source* package ‘ggstatsplot’ ...
** using staged installation
** R
** data
*** moving datasets to lazyload DB
** inst
** byte-compile and prepare package for lazy loading
Error in loadNamespace(j <- i[[1L]], c(lib.loc, .libPaths()), versionCheck = vI[[j]]) :
  namespace ‘effectsize’ 0.4.4-1 is being loaded, but >= 0.4.4.2 is required
Calls: <Anonymous> ... namespaceImportFrom -> asNamespace -> loadNamespace
Execution halted
ERROR: lazy loading failed for package ‘ggstatsplot’
* removing ‘/Library/Frameworks/R.framework/Versions/4.0/Resources/library/ggstatsplot’
* restoring previous ‘/Library/Frameworks/R.framework/Versions/4.0/Resources/library/ggstatsplot’
Error: Failed to install 'ggstatsplot' from GitHub:
  (converted from warning) installation of package ‘/var/folders/yv/69r1bw2931jb9tkdckxsmkr40000gn/T//Rtmp7q0MRR/file1516267a3e51b/ggstatsplot_0.8.0.tar.gz’ had non-zero exit status
```

I encountered a similar issue with the other installation method using remotes.

Cross linking to the issue: https://github.com/openjournals/joss-reviews/issues/3167
