### 2021-05-21 IndrajeetPatil/ggstatsplot#597 @njtierney
JOSS Review: Functionality
I encountered this error a few times:

```
Error in grid.Call(C_convert, x, as.integer(whatfrom), as.integer(whatto),  :
  Viewport has zero dimension(s)
```

This was solved by increasing the size of my RStudio viewer pane.

I noted that this was discussed in [this issue](https://github.com/IndrajeetPatil/ggstatsplot/issues/577), however the FAQ linked did not provide a suggestion on a way to improve this by resizing the plot

Cross linking to the JOSS review: https://github.com/openjournals/joss-reviews/issues/3167
