### 2021-05-21 IndrajeetPatil/ggstatsplot#598 @njtierney
JOSS Review: Benefits/Software implementation
In the paper, in the benefits section, it states that `ggstatsplot`:

> "increases reproducibility of statistical reporting"

I am not sure I agree with this fully, as while the package does an incredible job at presenting a high quality, readable graphics, it is not possible to replicate the statistical analysis separately.

As far as I can tell, the output is a graphic, and to reproduce just the analysis, not the graphic, would require inspecting either the source code of the functions, or manually reading and copying the details to a statistical test, which I think is prone to human error.

I think that the best way around this would be to either provide separate functions that the user could call, or for the plotting functions to perhaps provide an option where the statistical summary objects are provided as a list.

Cross linking to the review: https://github.com/openjournals/joss-reviews/issues/3167
