### 2021-05-21 #3167 @njtierney
Hi @csoneson ,

Thanks so much for your patience,

I've finished up the review now, had a few family/personal things come up which means I had reduced scope for other work.

Linked above are some issues that I noticed reading through the paper and the software. In particular, https://github.com/IndrajeetPatil/ggstatsplot/issues/598 I that it is hard to reproduce the statistical analysis without showing the graphic, I would be interested to discuss this and some solutions to this issue.
