### 2021-05-21 IndrajeetPatil/ggstatsplot#597 @IndrajeetPatil
Thanks! I've added this solution to the FAQ section now.
