### 2021-05-21 IndrajeetPatil/ggstatsplot#598 @IndrajeetPatil
> As far as I can tell, the output is a graphic, and to reproduce just the analysis, not the graphic, would require inspecting either the source code of the functions, or manually reading and copying the details to a statistical test, which I think is prone to human error.

This is true.

The `ggstatsplot` package is actually just a graphics package, while all the statistical analysis is carried out by the backend `statsExpressions` package (https://joss.theoj.org/papers/10.21105/joss.03236).

So, any statistical analysis in `ggstatsplot` can be reproduced with `statsExpressions` package:

``` r
options(tibble.width = Inf) # to see all columns

suppressPackageStartupMessages(library(ggstatsplot))
set.seed(123)
ggbetweenstats(mtcars, am, wt)
```

![](https://i.imgur.com/YZ3LpLw.png)

``` r
library(statsExpressions)
set.seed(123)
two_sample_test(mtcars, am, wt)
#> # A tibble: 1 x 16
#>   term  group mean.group1 mean.group2 statistic df.error    p.value
#>   <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>      <dbl>
#> 1 wt    am           3.77        2.41      5.49     29.2 0.00000627
#>   method                  estimate conf.level conf.low conf.high effectsize
#>   <chr>                      <dbl>      <dbl>    <dbl>     <dbl> <chr>
#> 1 Welch Two Sample t-test     1.88       0.95     1.05      2.70 Hedges' g
#>   conf.method conf.distribution expression
#>   <chr>       <chr>             <list>
#> 1 ncp         t                 <language>

two_sample_test(mtcars, am, wt, type = "bayesian")
#> # A tibble: 2 x 13
#>   term       estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference    -1.26       0.95   -1.79     -0.722     1               0
#> 2 Cohens_d       1.72       0.95    0.831     2.56      1               0
#>   prior.distribution prior.location prior.scale  bf10 method          expression
#>   <chr>                       <dbl>       <dbl> <dbl> <chr>           <list>
#> 1 cauchy                          0       0.707 1383. Bayesian t-test <language>
#> 2 cauchy                          0       0.707 1383. Bayesian t-test <language>
```

Additionally, the user can always access dataframes containing all the statistical analyses from the plot environment, although this is definitely not a beginner-friendly solution:

``` r
options(tibble.width = Inf) # to see all columns

suppressPackageStartupMessages(library(ggstatsplot))
set.seed(123)
p <- ggbetweenstats(mtcars, am, wt)

p$plot_env$subtitle_df
#> # A tibble: 1 x 16
#>   term  group mean.group1 mean.group2 statistic df.error    p.value
#>   <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>      <dbl>
#> 1 wt    am           3.77        2.41      5.49     29.2 0.00000627
#>   method                  estimate conf.level conf.low conf.high effectsize
#>   <chr>                      <dbl>      <dbl>    <dbl>     <dbl> <chr>
#> 1 Welch Two Sample t-test     1.88       0.95     1.05      2.70 Hedges' g
#>   conf.method conf.distribution expression
#>   <chr>       <chr>             <list>
#> 1 ncp         t                 <language>

p$plot_env$caption_df
#> # A tibble: 2 x 13
#>   term       estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference    -1.26       0.95   -1.79     -0.722     1               0
#> 2 Cohens_d       1.72       0.95    0.831     2.56      1               0
#>   prior.distribution prior.location prior.scale  bf10 method          expression
#>   <chr>                       <dbl>       <dbl> <dbl> <chr>           <list>
#> 1 cauchy                          0       0.707 1383. Bayesian t-test <language>
#> 2 cauchy                          0       0.707 1383. Bayesian t-test <language>
```

Do you advise that I be more explicit about `statsExpressions`'s role in statistical analysis?
Or would you much rather have me just remove the sentence in question?
