+++ tests/testthat/_snaps/ggcorrmat.md
@@ -648,12 +648,12 @@
       3       59 5.86e- 3 Winsorized Pearson correlation    61
       
       [[7]]
-      # A tibble: 3 x 15
-        parameter1 parameter2 estimate conf.level conf.low conf.high std.dev    pd
-        <chr>      <chr>         <dbl>      <dbl>    <dbl>     <dbl>   <dbl> <dbl>
-      1 brainwt    sleep_rem   -0.0908       0.95   -0.306     0.147   0.141 0.740
-      2 brainwt    bodywt       0.454        0.95    0.267     0.623   0.113 1    
-      3 sleep_rem  bodywt      -0.0956       0.95   -0.322     0.116   0.139 0.756
+      # A tibble: 3 x 14
+        parameter1 parameter2 estimate conf.level conf.low conf.high    pd
+        <chr>      <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>
+      1 brainwt    sleep_rem   -0.0911       0.95   -0.306     0.147 0.740
+      2 brainwt    bodywt       0.461        0.95    0.267     0.623 1    
+      3 sleep_rem  bodywt      -0.0959       0.95   -0.322     0.116 0.756
         rope.percentage prior.distribution prior.location prior.scale bayes.factor
                   <dbl> <chr>                       <dbl>       <dbl>        <dbl>
       1         0.430   beta                         1.41        1.41        0.269
@@ -666,12 +666,12 @@
       3 Bayesian Pearson correlation    48
       
       [[8]]
-      # A tibble: 3 x 15
-        parameter1 parameter2 estimate conf.level conf.low conf.high std.dev    pd
-        <chr>      <chr>         <dbl>      <dbl>    <dbl>     <dbl>   <dbl> <dbl>
-      1 brainwt    sleep_rem    -0.201       0.95   -0.402    0.0248  0.135  0.928
-      2 brainwt    bodywt        0.924       0.95    0.894    0.955   0.0203 1    
-      3 sleep_rem  bodywt       -0.306       0.95   -0.481   -0.129   0.114  0.990
+      # A tibble: 3 x 14
+        parameter1 parameter2 estimate conf.level conf.low conf.high    pd
+        <chr>      <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>
+      1 brainwt    sleep_rem    -0.205       0.95   -0.402    0.0248 0.928
+      2 brainwt    bodywt        0.926       0.95    0.894    0.955  1    
+      3 sleep_rem  bodywt       -0.310       0.95   -0.481   -0.129  0.990
         rope.percentage prior.distribution prior.location prior.scale bayes.factor
                   <dbl> <chr>                       <dbl>       <dbl>        <dbl>
       1          0.212  beta                         1.41        1.41     6.54e- 1
