### 2021-05-21 #3167 @IndrajeetPatil
@csoneson Can you please set the version of the package to `0.7.2` in this issue? I can't do it.

There was a new release of the package since the review started, which changes nothing about the paper but better for the record to be up-to-date.
