### 2021-05-21 #3167 @IndrajeetPatil
@njtierney Thanks for your comments!

> In particular, IndrajeetPatil/ggstatsplot#598 I that it is hard to reproduce the statistical analysis without showing the graphic, I would be interested to discuss this and some solutions to this issue.

I have proposed a couple of solutions in the related issue to address this, and I am happy to discuss this further.
