+++ appveyor.yml
@@ -34,15 +34,15 @@ environment:
     GCC_PATH: mingw_64
     PKGTYPE: both
 
-  - R_VERSION: devel
-    R_ARCH: x64
-    GCC_PATH: mingw_64
-    PKGTYPE: both
-
-  # - R_VERSION: oldrel
+  # - R_VERSION: devel
   #   R_ARCH: x64
   #   GCC_PATH: mingw_64
   #   PKGTYPE: both
+
+  - R_VERSION: oldrel
+    R_ARCH: x64
+    GCC_PATH: mingw_64
+    PKGTYPE: both
 
 matrix:
   fast_finish: true
