### 2021-05-21 IndrajeetPatil/ggstatsplot#596 @njtierney
Yup, sounds good to me! Just wanted to flag it as an issue in case it wasn't on your radar :)
