### 2021-05-21 IndrajeetPatil/ggstatsplot#598 @njtierney
This is great information. What do you think about a wrapper function like, `extract_ggstats_method()` or similar, to help extract this information from a plot?, which could do something like the following:

```r
extract_ggstats_method <- function(x, ...){
  stopifnot(inherits("gg"), "input must be a ggplot")
  list(substitle_info = x$plot_env$subtitle_df, caption_info = x$plot_env$caption_df)
}
```

With the documentation describing how to understand which test to run from `statsExpressions` based on this information, or cross linking to another page where you outline which tests are conducted in which functions.
