### 2021-05-22 dbd9155
closes #598; new extract_stats function
