+++ tests/testthat/test-extract_stats.R
@@ -1,0 +1,16 @@
+test_that(
+  desc = "checking if extract_stats works",
+  code = {
+    options(tibble.width = Inf)
+
+    set.seed(123)
+    p1 <- ggbetweenstats(mtcars, am, mpg)
+    expect_snapshot(extract_stats(p1))
+
+    set.seed(123)
+    p2 <- ggscatterstats(mtcars, wt, mpg, marginal = FALSE, type = "r")
+    expect_snapshot(extract_stats(p2))
+
+    expect_error(extract_stats(iris))
+  }
+)
