+++ man/ggbetweenstats.Rd
@@ -254,11 +254,7 @@ set.seed(123)
 library(ggstatsplot)
 
 # simple function call with the defaults
-ggbetweenstats(
-  data = mtcars,
-  x = am,
-  y = mpg
-)
+ggbetweenstats(mtcars, am, mpg)
 
 # more detailed function call
 ggbetweenstats(
