+++ NAMESPACE
@@ -8,6 +8,7 @@ export(as_tibble)
 export(combine_plots)
 export(effectsize)
 export(enframe)
+export(extract_stats)
 export(ggbarstats)
 export(ggbetweenstats)
 export(ggcoefstats)
