### 2021-05-22 IndrajeetPatil/ggstatsplot#598 @IndrajeetPatil
Thanks for the great suggestion!

I have implemented this function now: https://indrajeetpatil.github.io/ggstatsplot/reference/extract_stats.html

Example:

```r
set.seed(123)
library(ggstatsplot)
p <- ggbetweenstats(mtcars, am, mpg)
extract_stats(p)
#> $subtitle_data
#> # A tibble: 1 x 16
#>   term  group mean.group1 mean.group2 statistic df.error p.value
#>   <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>   <dbl>
#> 1 mpg   am           17.1        24.4     -3.77     18.3 0.00137
#>   method                  estimate conf.level conf.low conf.high effectsize
#>   <chr>                      <dbl>      <dbl>    <dbl>     <dbl> <chr>
#> 1 Welch Two Sample t-test    -1.35       0.95    -2.17    -0.512 Hedges' g
#>   conf.method conf.distribution expression
#>   <chr>       <chr>             <list>
#> 1 ncp         t                 <language>
#>
#> $caption_data
#> # A tibble: 2 x 13
#>   term       estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference     6.44       0.95     2.68    10.0   0.999               0
#> 2 Cohens_d      -1.30       0.95    -2.11    -0.519 0.999               0
#>   prior.distribution prior.location prior.scale  bf10 method          expression
#>   <chr>                       <dbl>       <dbl> <dbl> <chr>           <list>
#> 1 cauchy                          0       0.707  86.6 Bayesian t-test <language>
#> 2 cauchy                          0       0.707  86.6 Bayesian t-test <language>
#>
```

Let me know if you want me to make any other changes.
