+++ vignettes/web_only/combine_plots.Rmd
@@ -40,10 +40,16 @@ Lifecycle: [![lifecycle](https://img.shields.io/badge/lifecycle-stable-green.svg
 
 # Combining plots with `combine_plots`
 
-`ggstatsplot` contains a "helper" function named `combine_plots` to help you
-combine several plots into one plot or add a combination of title, caption, and
-annotation texts with suitable default parameters. It is a wrapper around
-`patchwork::wrap_plots()`. Some examples below.
+The full power of `ggstatsplot` can be leveraged with a functional programming
+package like [`purrr`](https://purrr.tidyverse.org/) that replaces `for` loops
+with code that is both more succinct and easier to read and, therefore, `purrr`
+should be preferrred `r emo::ji("heart_eyes_cat")`.
+
+In such cases, `ggstatsplot` contains a helper function `combine_plots` to
+combine multiple plots, which can be useful for combining a list of plots
+produced with `purrr`. This is a wrapper around `patchwork::wrap_plots` and lets
+you combine multiple plots and add a combination of title, caption, and
+annotation texts with suitable defaults.
 
 **Note Before**: If you have just one grouping variable and you'd like a plot
 for each `factor` of this variable the `grouped_` variants
