### 2021-05-22 IndrajeetPatil/ggstatsplot#598 @njtierney
That looks great to me, all I could think to suggest would be to add examples to the README and add tests, but I see you've already done that! Thanks for integrating the changes so quickly!
