### 2021-05-22 5bce218
retire ggstatsplot layer
