+++ R/ggbetweenstats.R
@@ -191,8 +191,7 @@ ggbetweenstats <- function(data,
                            ),
                            violin.args = list(width = 0.5, alpha = 0.2),
                            ggsignif.args = list(textsize = 3, tip_length = 0.01),
-                           ggtheme = ggplot2::theme_bw(),
-                           ggstatsplot.layer = TRUE,
+                           ggtheme = ggstatsplot::theme_ggstatsplot(),
                            package = "RColorBrewer",
                            palette = "Dark2",
                            ggplot.component = NULL,
@@ -441,7 +440,6 @@ ggbetweenstats <- function(data,
     subtitle = subtitle,
     caption = caption,
     ggtheme = ggtheme,
-    ggstatsplot.layer = ggstatsplot.layer,
     package = package,
     palette = palette,
     ggplot.component = ggplot.component
