+++ R/ggwithinstats.R
@@ -101,8 +101,7 @@ ggwithinstats <- function(data,
                           outlier.label.args = list(size = 3),
                           violin.args = list(width = 0.5, alpha = 0.2),
                           ggsignif.args = list(textsize = 3, tip_length = 0.01),
-                          ggtheme = ggplot2::theme_bw(),
-                          ggstatsplot.layer = TRUE,
+                          ggtheme = ggstatsplot::theme_ggstatsplot(),
                           package = "RColorBrewer",
                           palette = "Dark2",
                           ggplot.component = NULL,
@@ -315,7 +314,6 @@ ggwithinstats <- function(data,
     subtitle = subtitle,
     caption = caption,
     ggtheme = ggtheme,
-    ggstatsplot.layer = ggstatsplot.layer,
     package = package,
     palette = palette,
     ggplot.component = ggplot.component
