+++ DESCRIPTION
@@ -34,7 +34,7 @@ Imports:
     ipmisc (>= 6.0.2),
     pairwiseComparisons (>= 3.1.5),
     paletteer,
-    parameters (>= 0.13.0.1),
+    parameters (>= 0.13.0),
     patchwork,
     performance,
     purrr,
