+++ appveyor.yml
@@ -29,10 +29,10 @@ environment:
     #R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
 
   matrix:
-  # - R_VERSION: release
-  #   R_ARCH: x64
-  #   GCC_PATH: mingw_64
-  #   PKGTYPE: both
+  - R_VERSION: release
+    R_ARCH: x64
+    GCC_PATH: mingw_64
+    PKGTYPE: both
 
   # - R_VERSION: devel
   #   R_ARCH: x64
