### 2021-05-23 #3167 @IndrajeetPatil
Thanks, @csoneson.

I think this note is incorrect though.

> MISSING DOIs
> -  10.1016/0003-2697(86)90183-1 may be a valid DOI for title: The Elements of Graphing Data

The book in question was published in 1994, way before when `doi`s were introduced (2000). At least, I find no evidence that the suggested `doi` is for this book.
