### 2021-05-23 #3167 @csoneson
Yes, I agree - that `doi` appears to point to a review of the book.
