+++ .travis.yml
@@ -3,6 +3,7 @@ r:
   - release
   - devel
   - oldrel
+  - 3.6
 latex: true
 warnings_are_errors: TRUE
 
