+++ .github/workflows/check-full.yaml
@@ -29,7 +29,7 @@ jobs:
           - {os: ubuntu-16.04, r: 'release', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest", args: "--no-manual"}
           - {os: ubuntu-16.04, r: 'devel', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest", args: "--no-manual"}
           - {os: ubuntu-16.04, r: 'oldrel', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest", args: "--no-manual"}
-          - {os: ubuntu-16.04,   r: '3.6.0', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", args: "--no-manual"}
+          - {os: ubuntu-16.04, r: '3.6.0', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", args: "--no-manual"}
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
