+++ appveyor.yml
@@ -26,7 +26,7 @@ environment:
     NOT_CRAN: true
     _R_CHECK_FORCE_SUGGESTS: true
     _R_CHECK_CRAN_INCOMING_: true
-    #R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+    R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
 
   matrix:
   - R_VERSION: release
