### 2021-05-23 e66057e
simplify ggcorrmat a bit
