+++ tests/testthat/test-extract_stats.R
@@ -11,6 +11,10 @@ test_that(
     p2 <- ggscatterstats(mtcars, wt, mpg, marginal = FALSE, type = "r")
     expect_snapshot(extract_stats(p2))
 
+    set.seed(123)
+    p3 <- ggcorrmat(iris)
+    expect_snapshot(extract_stats(p3))
+
     expect_error(extract_stats(iris))
   }
 )
