+++ tests/testthat/test-ggcorrmat.R
@@ -33,7 +33,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-    expect_snapshot(p$labels)
+    expect_snapshot(list(p$labels, pb$plot$plot_env$legend.title))
   }
 )
 
@@ -59,7 +59,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-    expect_snapshot(p$labels)
+    expect_snapshot(list(p$labels, pb$plot$plot_env$legend.title))
   }
 )
 
@@ -90,7 +90,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-    expect_snapshot(p$labels)
+    expect_snapshot(list(p$labels, pb$plot$plot_env$legend.title))
   }
 )
 
@@ -124,7 +124,7 @@ test_that(
     # check data
     set.seed(123)
     expect_snapshot(pb$data)
-    expect_snapshot(p$labels)
+    expect_snapshot(list(p$labels, pb$plot$plot_env$legend.title))
   }
 )
 
@@ -144,7 +144,7 @@ test_that(
     # check data
     set.seed(123)
     # expect_snapshot(pb$data)
-    expect_snapshot(p$labels)
+    expect_snapshot(list(p$labels, pb$plot$plot_env$legend.title))
   }
 )
 
@@ -155,9 +155,8 @@ test_that(
   code = {
     skip_on_os("windows")
     skip_on_cran()
-    options(tibble.width = Inf, tibble.print_max = 50)
+    options(tibble.width = Inf)
     skip_on_ci()
-    skip_on_appveyor()
     skip_on_travis()
 
     set.seed(123)
