+++ vignettes/web_only/purrr_examples.Rmd
@@ -19,8 +19,14 @@ vignette: >
 ---
 
 ```{r setup, include = FALSE}
-# show me all columns
-options(tibble.width = Inf, pillar.bold = TRUE, pillar.subtle_num = TRUE)
+# pretty tibble printing
+options(
+  tibble.width = Inf,
+  pillar.bold = TRUE,
+  pillar.neg = TRUE,
+  pillar.subtle_num = TRUE,
+  pillar.min_chars = Inf
+)
 
 knitr::opts_chunk$set(
   collapse = TRUE,
