+++ .github/workflows/check-full.yaml
@@ -23,9 +23,9 @@ jobs:
           - {os: windows-latest, r: 'release', args: "--no-manual"}
           - {os: windows-latest, r: 'devel', args: "--no-manual"}
           - {os: windows-latest, r: 'oldrel', args: "--no-manual"}
-          - {os: windows-latest, r: '3.6.0', args: "--no-manual"}
+          # 'Rmpfr' not available on 3.6 for Windows
           - {os: macOS-latest, r: 'release', args: "--no-manual"}
-          - {os: macOS-latest, r: 'devel', args: "--no-manual"}
+          #- {os: macOS-latest, r: 'devel', args: "--no-manual"}
           - {os: macOS-latest, r: 'oldrel', args: "--no-manual"}
           - {os: macOS-latest, r: '3.6.0', args: "--no-manual"}
           - {os: ubuntu-16.04, r: 'release', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest", args: "--no-manual"}
