+++ man/theme_ggstatsplot.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helpers_theme_ggstatsplot.R
+% Please edit documentation in R/theme_ggstatsplot.R
 \name{theme_ggstatsplot}
 \alias{theme_ggstatsplot}
 \title{Default theme used in all \code{ggstatsplot} package plots}
