### 2021-05-24 2e9c70f
Small fixes to paper and bibliography
