+++ paper/paper.bib
@@ -44,7 +44,7 @@
   isbn = {978-1-4338-0561-5},
   language = {English},
   publisher = {{American Psychological Association}},
-  author = {American Psychological Association},
+  author = {{American Psychological Association}},
   year = {2019},
 }
 
@@ -117,7 +117,7 @@
   }
 
 @Article{Makowski2019,
-    title = {bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework.},
+    title = {bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the {B}ayesian Framework.},
     author = {Dominique Makowski and Mattan S. Ben-Shachar and Daniel Lüdecke},
     journal = {Journal of Open Source Software},
     doi = {10.21105/joss.01541},
@@ -151,7 +151,7 @@
 
 
 @article{van2020jasp,
-  title={The JASP guidelines for conducting and reporting a Bayesian analysis},
+  title={The JASP guidelines for conducting and reporting a {B}ayesian analysis},
   author={van Doorn, Johnny and van den Bergh, Don and B{\"o}hm, Udo and Dablander, Fabian and Derks, Koen and Draws, Tim and Etz, Alexander and Evans, Nathan J and Gronau, Quentin F and Haaf, Julia M and others},
   journal={Psychonomic Bulletin \& Review},
   pages={1--14},
@@ -185,7 +185,7 @@
   title={{R for Data Science}},
   author={Wickham, Hadley and Grolemund, Garrett},
   year={2016},
-  publisher={O'Reilly Medias}
+  publisher={O'Reilly Media}
 }
 
 @book{grant2018data,
