### 2021-05-24 #3167 @IndrajeetPatil
- [ ] add all the authors for Doorn et al
- [ ] add ISBNs to all the books for consistency
- [ ] fix the titles of Ludecke et al (2020, 2021) - the ones given in the bibliography are not the same as the papers that the DOIs point to
- [ ] replace the statsExpressions preprint with the link to the JOSS paper (I saw that it was accepted recently)

Once this is done, the next steps are as follows:

- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the Zenodo archive has the same title and author list as the paper. You may also add the authors' ORCID.
