### 2021-05-24 1852987
address @csoneson comments for JOSS
