+++ paper/paper.bib
@@ -30,7 +30,6 @@
   edition = {1st edition},
   title = {The {{Elements}} of {{Graphing Data}}},
   isbn = {978-0-534-03729-1},
-  abstract = {--This text refers to an out of print or unavailable edition of this title.},
   language = {English},
   publisher = {{Wadsworth, Inc.}},
   author = {William S. Cleveland},
@@ -60,36 +59,36 @@
   }
 
 @incollection{Wilkinson2012,
-  title={The grammar of graphics},
-  author={Wilkinson, Leland},
-  booktitle={Handbook of computational statistics},
-  pages={375--414},
-  year={2012},
+  title = {{The Grammar of Graphics}},
+  author = {Wilkinson, Leland},
+  booktitle = {Handbook of computational statistics},
+  pages = {375--414},
+  year = {2012},
+  isbn = {978-3540404644},
   publisher={Springer}
 }
+
 @Book{Wickham2016,
     author = {Hadley Wickham},
     title = {{ggplot2}: Elegant Graphics for Data Analysis},
     publisher = {Springer-Verlag New York},
     year = {2016},
-    isbn = {978-3-319-24277-4},
-    url = {https://ggplot2.tidyverse.org},
+    isbn = {978-3-319-24277-4}
   }
-
-@book{chambers2019seven,
-  title={The seven deadly sins of psychology: A manifesto for reforming the culture of scientific practice},
-  author={Chambers, Chris},
-  year={2019},
-  publisher={Princeton University Press}
-}
 
 @Article{Patil2021,
-    title = {{statsExpressions}: R Package for Tidy Dataframes and Expressions with Statistical Details},
-    author = {Indrajeet Patil},
+    doi = {10.21105/joss.03236},
+    url = {https://doi.org/10.21105/joss.03236},
     year = {2021},
-    journal = {PsyArXiv},
-    doi = {10.31234/osf.io/ntbvy},
+    publisher = {{The Open Journal}},
+    volume = {6},
+    number = {61},
+    pages = {3236},
+    author = {{Indrajeet Patil}},
+    title = {{statsExpressions: {R} Package for Tidy Dataframes and Expressions with Statistical Details}},
+    journal = {{Journal of Open Source Software}},
   }
+
 
 @Article{Ben-Shachar2020,
     title = {{e}ffectsize: Estimation of Effect Size Indices and Standardized Parameters},
@@ -106,7 +105,7 @@
 
 
 @Article{Lüdecke2020parameters,
-    title = {parameters: Extracting, Computing and Exploring the Parameters of Statistical Models using {R}.},
+    title = {Extracting, Computing and Exploring the Parameters of Statistical Models using {R}},
     volume = {5},
     doi = {10.21105/joss.02445},
     number = {53},
@@ -128,15 +127,16 @@
     url = {https://joss.theoj.org/papers/10.21105/joss.01541},
   }
 
-
 @Article{Lüdecke2020performance,
-    title = {Assessment, Testing and Comparison of Statistical Models using {R}},
+    title = {{performance}: An {R} Package for Assessment, Comparison and Testing of Statistical Models},
     author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Philip Waggoner and Dominique Makowski},
     year = {2021},
-    journal = {PsyArXiv},
-    doi = {10.31234/osf.io/vtq8f},
+    journal = {Journal of Open Source Software},
+    volume = {6},
+    number = {60},
+    pages = {3139},
+    doi = {10.21105/joss.03139},
   }
-
 
[…]
