+++ paper/paper.bib
@@ -84,7 +84,7 @@
     volume = {6},
     number = {61},
     pages = {3236},
-    author = {{Indrajeet Patil}},
+    author = {Indrajeet Patil},
     title = {{statsExpressions: {R} Package for Tidy Dataframes and Expressions with Statistical Details}},
     journal = {{Journal of Open Source Software}},
   }
@@ -152,26 +152,26 @@
 
 @article{van2020jasp,
   title={The JASP guidelines for conducting and reporting a {B}ayesian analysis},
-  author={van Doorn and Johnny
-  van den Bergh and Don
-  Böhm and Udo
-  Dablander and Fabian
-  Derks and Koen
-  Draws and Tim
-  Etz and Alexander
-  Evans and Nathan J
-  Gronau and Quentin F
-  Haaf and Julia M
-  Hinne and Max
-  Kucharský and Šimon
-  Ly and Alexander
-  Marsman and Maarten
-  Matzke and Dora
-  Gupta and Akash R Komarlu Narendra
-  Sarafoglou and Alexandra
-  Stefan and Angelika
-  Voelkel and Jan G
-  Wagenmakers and Eric-Jan},
+  author={Johnny van Doorn and
+  Don van den Bergh and
+  Udo Böhm and
+  Fabian Dablander and
+  Koen Derks and
+  Tim Draws and
+  Alexander Etz and
+  Nathan J Evans and
+  Quentin F Gronau and
+  Julia M Haaf and
+  Max Hinne and
+  Šimon Kucharský and
+  AlexanderLy and
+  Maarten Marsman and
+  Dora Matzke and
+  Akash R Komarlu Narendra Gupta and
+  Alexandra Sarafoglou and
+  Angelika Stefan and
+  Jan G Voelkel and
+  Eric-Jan Wagenmakers},
   journal={Psychonomic Bulletin \& Review},
   pages={1--14},
   year={2020},
