### 2021-05-25 #3167 @arfon
> @openjournals/dev - any idea what might be going on here? I'm not sure the issue was introduced by the last round of fixes; right now I'm also not able to compile the version I forked before those changes were made (using the preview service linked in https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles), which I think should be the same version that whedon compiled a couple of days ago above.

It was the `paper.tex` file that was confusing Whedon. Looks like @IndrajeetPatil removed this and so the compilation now works.
