### 2021-05-25 #3167 @IndrajeetPatil
Thanks, @arfon!
Surprised that it's working, because I had tried to compile the paper using http://whedon.theoj.org/ after the commit that removed `.tex` file, but it was still failing.

But I am glad that it's working now! 😄
