### 2021-05-25 #3167 @csoneson
Excellent, thank you! Only one minor detail I noticed - in the Doorn et al, the first and second name of `Alexander Ly` have been merged, so they're not properly represented. Could you fix that? And since you're anyway modifying that entry, could you also put some braces around `van Doorn` and `van den Bergh`, so that they render nicely together. In the meantime, I'll log the archive and version info.
