+++ paper/paper.md
@@ -16,7 +16,7 @@ affiliations:
   - name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
     index: 1
 bibliography: paper.bib
-date: "2021-05-24"
+date: "2021-05-25"
 ---
 
 
