+++ paper/paper.bib
@@ -152,8 +152,8 @@
 
 @article{van2020jasp,
   title={The JASP guidelines for conducting and reporting a {B}ayesian analysis},
-  author={Johnny van Doorn and
-  Don van den Bergh and
+  author={Johnny {van Doorn} and
+  Don {van den Bergh} and
   Udo Böhm and
   Fabian Dablander and
   Koen Derks and
@@ -164,7 +164,7 @@
   Julia M Haaf and
   Max Hinne and
   Šimon Kucharský and
-  AlexanderLy and
+  Alexander Ly and
   Maarten Marsman and
   Dora Matzke and
   Akash R Komarlu Narendra Gupta and
