### 2021-05-25 #3167 @IndrajeetPatil
@csoneson Thanks for pointing this out!

Fixed in https://github.com/IndrajeetPatil/ggstatsplot/commit/8517c2c5a4613f63c99e70dd53c88d030895fd28
