### 2021-05-25 afc6ebb
add more details to extract_stats
