+++ tests/testthat/test-extract_stats.R
@@ -15,6 +15,18 @@ test_that(
     p3 <- ggcorrmat(iris)
     expect_snapshot(extract_stats(p3))
 
+    set.seed(123)
+    p4 <- ggbetweenstats(mtcars, cyl, mpg)
+    expect_snapshot(extract_stats(p4)$pairwise_comparisons_data)
+
+    set.seed(123)
+    p5 <- ggpiestats(mtcars, cyl)
+    expect_snapshot(extract_stats(p5))
+
+    set.seed(123)
+    p6 <- ggbarstats(mtcars, cyl, am)
+    expect_snapshot(extract_stats(p6))
+
     expect_error(extract_stats(iris))
   }
 )
