+++ R/extract_stats.R
@@ -11,6 +11,10 @@
 #' following vignette:
 #' https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html
 #'
+#' @return
+#'
+#' A list of tibbles containing statistical analysis summaries.
+#'
 #' @note
 #' Note that if you want to use this function with `ggscatterstats`, you will
 #' have to set `marginal = FALSE` to return an object of `ggplot` type.
@@ -22,8 +26,13 @@
 #' \donttest{
 #' set.seed(123)
 #' library(ggstatsplot)
-#' p <- ggbetweenstats(mtcars, am, mpg)
+#'
+#' # in case of group comparisons
+#' p <- ggbetweenstats(mtcars, cyl, mpg)
 #' extract_stats(p)
+#'
+#' # the exact details depend on the function
+#' extract_stats(ggbarstats(mtcars, cyl, am))
 #' }
 #' @export
 
@@ -31,9 +40,12 @@ extract_stats <- function(p, ...) {
   # check the input
   if (!inherits(p, "gg")) stop("Input must be a 'ggplot' object.", call. = FALSE)
 
-  # caption dataframe might not be present when `type` is not "parametric"
+  # the exact details will depend on the function
   list(
     subtitle_data = tryCatch(p$plot_env$subtitle_df, error = function(e) NULL),
-    caption_data = tryCatch(p$plot_env$caption_df, error = function(e) NULL)
+    caption_data = tryCatch(p$plot_env$caption_df, error = function(e) NULL),
+    pairwise_comparisons_data = tryCatch(p$plot_env$mpc_df, error = function(e) NULL),
+    descriptive_data = tryCatch(p$plot_env$descriptive_df, error = function(e) NULL),
+    one_sample_data = tryCatch(p$plot_env$onesample_df, error = function(e) NULL)
   )
 }
