+++ R/ggbetweenstats.R
@@ -400,7 +400,7 @@ ggbetweenstats <- function(data,
 
   if (isTRUE(pairwise.comparisons) && test == "anova") {
     # creating dataframe with pairwise comparison results
-    df_mcp <- pairwiseComparisons::pairwise_comparisons(
+    mpc_df <- pairwiseComparisons::pairwise_comparisons(
       data = data,
       x = {{ x }},
       y = {{ y }},
@@ -415,7 +415,7 @@ ggbetweenstats <- function(data,
     # adding the layer for pairwise comparisons
     plot <- ggsignif_adder(
       plot = plot,
-      df_mcp = df_mcp,
+      mpc_df = mpc_df,
       data = data,
       x = {{ x }},
       y = {{ y }},
@@ -427,7 +427,7 @@ ggbetweenstats <- function(data,
     if (type != "bayes") {
       caption <- pairwiseComparisons::pairwise_caption(
         caption,
-        unique(df_mcp$test.details),
+        unique(mpc_df$test.details),
         pairwise.display
       )
     }

[…]
